Split db.test.ts into focused files and drop tautologies
The 4048-line test/lib/db.test.ts had a handful of quality problems:

- Tautological assertions (e.g. "CONFIG_KEYS contains expected keys" just
  re-stated the constant, "LATEST_UPDATE constant is exported" checked that a
  string was a string).
- Two near-identical tests for the reserveSession stale-retry path.
- Duplicate table-helper tests split across sibling describes.
- A single file far larger than the project's 400-line guideline.

This commit reorganises those tests into one file per module under
test/lib/db/ (attendees, auth, events, groups, migrations, settings,
sessions, processed-payments, legacy-migration, etc.), removes the
duplicates and tautologies, and promotes getTestPrivateKey into #test-utils
so the decryption helper isn't copy-pasted in each split file.

Also adds a guide FAQ describing what changing an admin password does
(session invalidation + DATA_KEY re-wrap keeps old attendee data readable),
reflecting behaviour already covered by the auth tests but previously
undocumented.
