--- test/lib/db.test.ts
@@ -1,4048 +1,0 @@
-import { createClient, type ResultSet } from "@libsql/client";
-import { expect } from "@std/expect";
-import { afterEach, beforeEach, describe, it as test } from "@std/testing/bdd";
-import { spy, stub } from "@std/testing/mock";
-import { FakeTime } from "@std/testing/time";
-import { decryptWithKey } from "#lib/crypto/encryption.ts";
-import { deriveKEK, importPrivateKey, unwrapKey } from "#lib/crypto/keys.ts";
-import {
-  getAllActivityLog,
-  getEventActivityLog,
-  getEventWithActivityLog,
-  logActivity,
-} from "#lib/db/activityLog.ts";
-import {
-  checkBatchAvailability,
-  createAttendeeAtomic,
-  decryptAttendees,
-  deleteAttendee,
-  getActiveEventStats,
-  getAttendee,
-  getAttendeesByTokens,
-  getAttendeesRaw,
-  getDateAttendeeCount,
-  getNewestAttendeesRaw,
-  hasAvailableSpots,
-  updateCheckedIn,
-} from "#lib/db/attendees.ts";
-import { getDb, insert, rawSql, setDb } from "#lib/db/client.ts";
-import {
-  computeSlugIndex,
-  deleteEvent,
-  eventsTable,
-  getAllEvents,
-  getEvent,
-  getEventsBySlugsBatch,
-  getEventWithAttendeeRaw,
-  getEventWithAttendeesRaw,
-  getEventWithCount,
-  isSlugTaken,
-  writeClosesAt,
-  writeEventDate,
-} from "#lib/db/events.ts";
-import {
-  computeGroupSlugIndex,
-  getActiveEventsByGroupId,
-  getAllGroups,
-  getGroupBySlugIndex,
-  groupsTable,
-  isGroupSlugTaken,
-  resetGroupEvents,
-} from "#lib/db/groups.ts";
-import {
-  clearLoginAttempts,
-  isLoginRateLimited,
-  recordFailedLogin,
-} from "#lib/db/login-attempts.ts";
-import {
-  initDb,
-  LATEST_UPDATE,
-  resetDatabase,
-  SCHEMA_HASH,
-} from "#lib/db/migrations.ts";
-import {
-  finalizeSession as finalizePaymentSession,
-  isSessionProcessed,
-  reserveSession,
-  STALE_RESERVATION_MS,
-} from "#lib/db/processed-payments.ts";
-import {
-  createSession,
-  deleteAllSessions,
-  deleteOtherSessions,
-  deleteSession,
-  getAllSessions,
-  getSession,
-} from "#lib/db/sessions.ts";
-import { CONFIG_KEYS, settings } from "#lib/db/settings.ts";
-import { getUserByUsername, verifyUserPassword } from "#lib/db/users.ts";
-import { nowMs } from "#lib/now.ts";
-import {
-  createPaidTestAttendee,
-  createTestAttendee,
-  createTestEvent,
-  createTestGroup,
-  describeWithEnv,
-  invalidateTestDbCache,
-  resetDb,
-  setTestEnv,
-  setupTestEncryptionKey,
-  TEST_ADMIN_PASSWORD,
-  TEST_ADMIN_USERNAME,
-} from "#test-utils";
-
-/** Helper to get private key for decrypting attendees in tests */
-const getTestPrivateKey = async (): Promise<CryptoKey> => {
-  const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-  if (!user) throw new Error("Test setup failed: user not found");
-  const passwordHash = await verifyUserPassword(user, TEST_ADMIN_PASSWORD);
-  if (!passwordHash) throw new Error("Test setup failed: invalid password");
-  if (!user.wrapped_data_key)
-    throw new Error("Test setup failed: no wrapped data key");
-  const kek = await deriveKEK(passwordHash);
-  const dataKey = await unwrapKey(user.wrapped_data_key, kek);
-  const wrappedPrivateKey = settings.wrappedPrivateKey;
-  if (!wrappedPrivateKey)
-    throw new Error("Test setup failed: no wrapped private key");
-  const privateKeyJwk = await decryptWithKey(wrappedPrivateKey, dataKey);
-  return importPrivateKey(privateKeyJwk);
-};
-
-describeWithEnv("db", { db: true }, () => {
-  describe("getDb", () => {
-    test("throws error when DB_URL is not set", () => {
-      setDb(null);
-      const restore = setTestEnv({ DB_URL: undefined });
-      try {
-        expect(() => getDb()).toThrow(
-          "DB_URL environment variable is required",
-        );
-      } finally {
-        restore();
-      }
-    });
-  });
-
-  describe("insert()", () => {
-    test("builds sql and args from record", () => {
-      const stmt = insert("users", {
-        name: "Alice",
-        email: "a@b.com",
-      });
-      expect(stmt.sql).toBe(
-        "INSERT INTO users (name, email)" + " VALUES (?, ?)",
-      );
-      expect(stmt.args).toEqual(["Alice", "a@b.com"]);
-    });
-
-    test("supports rawSql for expressions", () => {
-      const stmt = insert("event_attendees", {
-        event_id: 1,
-        attendee_id: rawSql("last_insert_rowid()"),
-        quantity: 2,
-      });
-      expect(stmt.sql).toBe(
-        "INSERT INTO event_attendees" +
-          " (event_id, attendee_id, quantity)" +
-          " VALUES (?, last_insert_rowid(), ?)",
-      );
-      expect(stmt.args).toEqual([1, 2]);
-    });
-
-    test("handles null values as params", () => {
-      const stmt = insert("payments", {
-        id: "p1",
-        attendee_id: null,
-        created: "now",
-      });
-      expect(stmt.sql).toBe(
-        "INSERT INTO payments" +
-          " (id, attendee_id, created)" +
-          " VALUES (?, ?, ?)",
-      );
-      expect(stmt.args).toEqual(["p1", null, "now"]);
-    });
-
-    test("executes correctly against db", async () => {
-      const stmt = insert("settings", {
-        key: "insert_test",
-        value: "works",
-      });
-      await getDb().execute(stmt);
-      const row = await getDb().execute(
-        "SELECT value FROM settings" + " WHERE key = 'insert_test'",
-      );
-      expect(row.rows[0]!.value).toBe("works");
-    });
-  });
-
-  describe("initDb version check", () => {
-    test("LATEST_UPDATE constant is exported", () => {
-      expect(typeof LATEST_UPDATE).toBe("string");
-      expect(LATEST_UPDATE.length).toBeGreaterThan(0);
-    });
-
-    test("initDb stores latest_db_update in settings", async () => {
-      const result = await getDb().execute(
-        "SELECT value FROM settings WHERE key = 'latest_db_update'",
-      );
-      expect(result.rows[0]?.value).toBe(LATEST_UPDATE);
-    });
-
-    test("initDb stores db_schema_hash in settings", async () => {
-      const result = await getDb().execute(
-        "SELECT value FROM settings WHERE key = 'db_schema_hash'",
-      );
-      expect(result.rows[0]?.value).toBe(SCHEMA_HASH);
-    });
-
-    test("initDb re-runs when schema hash changes", async () => {
-      // Simulate a stale hash (schema changed but LATEST_UPDATE wasn't bumped)
-      await getDb().execute(
-        "UPDATE settings SET value = 'stale' WHERE key = 'db_schema_hash'",
-      );
-
-      // initDb should detect the mismatch and re-run
-      await initDb();
-
-      // Verify hash was updated back to current
-      const result = await getDb().execute(
-        "SELECT value FROM settings WHERE key = 'db_schema_hash'",
-      );
-      expect(result.rows[0]?.value).toBe(SCHEMA_HASH);
-    });
-
-    test("initDb can be called multiple times safely", async () => {
-      // initDb was already called in beforeEach, call it again
-      await initDb();
-
-      // Verify tables still exist and work
-      const events = await getAllEvents();
-      expect(events).toEqual([]);
-    });
-
-    test("initDb bails early when database is up to date", async () => {
-      // initDb was already called and stored LATEST_UPDATE
-      // Calling it again should bail early without error
-      const startTime = performance.now();
-      await initDb();
-      const duration = performance.now() - startTime;
-
-      // Should be very fast since it bails early (typically < 5ms)
-      // We just check it completes without error
-      expect(duration).toBeLessThan(100);
-    });
-
-    test("initDb drops legacy indexes not in declarative schema", async () => {
-      // Create a legacy index that the declarative schema doesn't declare
-      await getDb().execute(
-        `CREATE INDEX IF NOT EXISTS
-         idx_attendees_legacy_created
-         ON attendees(created)`,
-      );
-
-      // Verify the legacy index exists
-      const before = await getDb().execute(
-        `SELECT name FROM sqlite_master
-         WHERE type = 'index'
-           AND name = 'idx_attendees_legacy_created'`,
-      );
-      expect(before.rows.length).toBe(1);
-
-      // Force re-run by invalidating schema hash
-      await getDb().execute(
-        "UPDATE settings SET value = 'stale' WHERE key = 'db_schema_hash'",
-      );
-      await initDb();
-
-      // Legacy index should be dropped
-      const after = await getDb().execute(
-        `SELECT name FROM sqlite_master
-         WHERE type = 'index'
-           AND name = 'idx_attendees_legacy_created'`,
-      );
-      expect(after.rows.length).toBe(0);
-    });
-  });
-
-  describe("pre-migration backup", () => {
-    test("creates backup before migrating existing database when storage is enabled", async () => {
-      const tmpDir = Deno.makeTempDirSync();
-      const restore = setTestEnv({ LOCAL_STORAGE_PATH: tmpDir });
-      try {
-        // Force re-run by invalidating schema hash
-        await getDb().execute(
-          "UPDATE settings SET value = 'stale' WHERE key = 'db_schema_hash'",
-        );
-        await initDb();
-
-        // A backup zip should have been written to the temp directory
-        const files = [...Deno.readDirSync(tmpDir)]
-          .map((e) => e.name)
-          .filter((n) => n.startsWith("backup-") && n.endsWith(".zip"));
-        expect(files.length).toBe(1);
-      } finally {
-        restore();
-        Deno.removeSync(tmpDir, { recursive: true });
-      }
-    });
-
-    test("skips backup on brand-new database", async () => {
-      const tmpDir = Deno.makeTempDirSync();
-      const restore = setTestEnv({ LOCAL_STORAGE_PATH: tmpDir });
-      try {
-        // Reset database completely and reinitialize
-        await resetDatabase();
-        invalidateTestDbCache();
-        await initDb();
-
-        // No backup should have been created — there was no existing data
-        const files = [...Deno.readDirSync(tmpDir)]
-          .map((e) => e.name)
-          .filter((n) => n.startsWith("backup-") && n.endsWith(".zip"));
-        expect(files.length).toBe(0);
-      } finally {
-        restore();
-        Deno.removeSync(tmpDir, { recursive: true });
-      }
-    });
-
-    test("skips backup when storage is not enabled", async () => {
-      const restore = setTestEnv({
-        LOCAL_STORAGE_PATH: undefined,
-        STORAGE_ZONE_NAME: undefined,
-        STORAGE_ZONE_KEY: undefined,
-      });
-      try {
-        await getDb().execute(
-          "UPDATE settings SET value = 'stale' WHERE key = 'db_schema_hash'",
-        );
-        await initDb();
-
-        // Migration completes — no backup attempted (no storage configured)
-        const result = await getDb().execute(
-          "SELECT value FROM settings WHERE key = 'db_schema_hash'",
-        );
-        expect(result.rows[0]?.value).toBe(SCHEMA_HASH);
-      } finally {
-        restore();
-      }
-    });
-
-    test("blocks migration when backup fails", async () => {
-      // Use Bunny CDN storage (no LOCAL_STORAGE_PATH) with dummy credentials
-      // so uploadRaw hits the network and fails
-      const restore = setTestEnv({
-        STORAGE_ZONE_NAME: "fake-zone",
-        STORAGE_ZONE_KEY: "fake-key",
-        LOCAL_STORAGE_PATH: undefined,
-      });
-      try {
-        await getDb().execute(
-          "UPDATE settings SET value = 'stale' WHERE key = 'db_schema_hash'",
-        );
-        await expect(initDb()).rejects.toThrow();
-
-        // Schema hash should still be stale — migration did not proceed
-        const result = await getDb().execute(
-          "SELECT value FROM settings WHERE key = 'db_schema_hash'",
-        );
-        expect(result.rows[0]?.value).toBe("stale");
-      } finally {
-        restore();
-        // Clean up the migration lock so subsequent tests aren't blocked
-        await getDb().execute(
-          "DELETE FROM settings WHERE key = 'migration_lock'",
-        );
-      }
-    });
-  });
-
-  describe("resetDatabase", () => {
-    test("drops all tables", async () => {
-      // Create some data first
-      await createTestEvent({
-        name: "Test Event",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      await createSession(
-        "test-token",
-        "test-csrf",
-        Date.now() + 1000,
-        null,
-        1,
-      );
-
-      // Reset the database
-      await resetDatabase();
-      invalidateTestDbCache();
-
-      // Verify tables are gone by checking that they don't exist
-      const client = getDb();
-
-      // Check that events table was dropped
-      const tablesResult = await client.execute(
-        "SELECT name FROM sqlite_master WHERE type='table'",
-      );
-      const tableNames = tablesResult.rows.map((r) => r.name);
-      expect(tableNames).not.toContain("events");
-      expect(tableNames).not.toContain("attendees");
-      expect(tableNames).not.toContain("sessions");
-      expect(tableNames).not.toContain("settings");
-      expect(tableNames).not.toContain("login_attempts");
-      expect(tableNames).not.toContain("processed_payments");
-      expect(tableNames).not.toContain("activity_log");
-    });
-
-    test("can reinitialize database after reset", async () => {
-      // Reset and reinitialize
-      await resetDatabase();
-      invalidateTestDbCache();
-      await initDb();
-
-      // Verify we can create new data
-      await settings.setup.complete("testadmin", TEST_ADMIN_PASSWORD, "USD");
-      const event = await createTestEvent({
-        name: "New Event",
-        maxAttendees: 25,
-        thankYouUrl: "https://example.com",
-      });
-
-      expect(event.id).toBe(1);
-      expect(event.name).toBe("New Event");
-    });
-  });
-
-  describe("settings", () => {
-    test("getSetting returns null for missing key", () => {
-      const value = settings.getCachedRaw("missing");
-      expect(value).toBeNull();
-    });
-
-    test("setSetting and getSetting work together", async () => {
-      await settings.setRaw("test_key", "test_value");
-      await settings.loadAll();
-      const value = settings.getCachedRaw("test_key");
-      expect(value).toBe("test_value");
-    });
-
-    test("setSetting overwrites existing value", async () => {
-      await settings.setRaw("key", "value1");
-      await settings.setRaw("key", "value2");
-      await settings.loadAll();
-      const value = settings.getCachedRaw("key");
-      expect(value).toBe("value2");
-    });
-  });
-
-  describe("buildSnapshot via loadAll", () => {
-    test("loads valid payment provider from raw settings", async () => {
-      await settings.setRaw("payment_provider", "stripe");
-      settings.invalidateCache();
-      await settings.loadAll();
-      expect(settings.paymentProvider).toBe("stripe");
-    });
-
-    test("ignores invalid payment provider in raw settings", async () => {
-      await settings.setRaw("payment_provider", "not-a-provider");
-      settings.invalidateCache();
-      await settings.loadAll();
-      expect(settings.paymentProvider).toBeNull();
-    });
-  });
-
-  describe("setup", () => {
-    test("completeSetup sets all config values and generates key hierarchy", async () => {
-      // Delete existing user from createTestDbWithSetup to test fresh setup
-      await getDb().execute("DELETE FROM users");
-      await getDb().execute("DELETE FROM settings");
-      await settings.setup.complete("setupuser", "mypassword", "US");
-      settings.invalidateCache();
-      await settings.loadAll();
-
-      expect(await settings.setup.isComplete()).toBe(true);
-      // Password is now stored on the user row, verify via user-based API
-      const user = await getUserByUsername("setupuser");
-      expect(user).not.toBeNull();
-      const hash = await verifyUserPassword(user!, "mypassword");
-      expect(hash).toBeTruthy();
-      expect(hash).toContain("pbkdf2:");
-      expect(settings.currency).toBe("USD");
-
-      // Key hierarchy should be generated
-      expect(settings.publicKey).toBeTruthy();
-      expect(user!.wrapped_data_key).toBeTruthy();
-      expect(settings.wrappedPrivateKey).toBeTruthy();
-    });
-
-    test("isComplete reloads cache when it has expired", async () => {
-      settings.invalidateCache();
-      // Cache is now invalid; isComplete should reload it from the DB
-      const result = await settings.setup.isComplete();
-      expect(result).toBe(true);
-    });
-
-    test("CONFIG_KEYS contains expected keys", () => {
-      expect(CONFIG_KEYS.COUNTRY).toBe("country");
-      expect(CONFIG_KEYS.SETUP_COMPLETE).toBe("setup_complete");
-      expect(CONFIG_KEYS.WRAPPED_PRIVATE_KEY).toBe("wrapped_private_key");
-      expect(CONFIG_KEYS.PUBLIC_KEY).toBe("public_key");
-      expect(CONFIG_KEYS.STRIPE_SECRET_KEY).toBe("stripe_secret_key");
-    });
-
-    test("getCurrencyCodeFromDb returns GBP by default", () => {
-      expect(settings.currency).toBe("GBP");
-    });
-
-    test("getCountryFromDb returns GB when no country is stored", async () => {
-      await getDb().execute("DELETE FROM settings");
-      settings.invalidateCache();
-      expect(settings.country).toBe("GB");
-    });
-  });
-
-  describe("stripe key", () => {
-    test("hasStripeKey returns false when not set", () => {
-      expect(settings.stripe.hasKey).toBe(false);
-    });
-
-    test("hasStripeKey returns true after setting key", async () => {
-      await settings.update.stripe.secretKey("sk_test_123");
-      expect(settings.stripe.hasKey).toBe(true);
-    });
-
-    test("getStripeSecretKeyFromDb returns empty string when not set", () => {
-      expect(settings.stripe.secretKey).toBe("");
-    });
-
-    test("getStripeSecretKeyFromDb returns decrypted key after setting", async () => {
-      await settings.update.stripe.secretKey("sk_test_secret_key");
-      const key = settings.stripe.secretKey;
-      expect(key).toBe("sk_test_secret_key");
-    });
-
-    test("updateStripeKey stores key encrypted", async () => {
-      await settings.update.stripe.secretKey("sk_test_encrypted");
-      await settings.loadAll();
-      // Verify the raw value in DB is encrypted (starts with enc:1:)
-      const rawValue = settings.getCachedRaw(CONFIG_KEYS.STRIPE_SECRET_KEY);
-      expect(rawValue).toMatch(/^enc:1:/);
-      // But getStripeSecretKeyFromDb returns decrypted
-      expect(settings.stripe.secretKey).toBe("sk_test_encrypted");
-    });
-
-    test("updateStripeKey overwrites existing key", async () => {
-      await settings.update.stripe.secretKey("sk_test_first");
-      expect(settings.stripe.secretKey).toBe("sk_test_first");
-
-      await settings.update.stripe.secretKey("sk_test_second");
-      expect(settings.stripe.secretKey).toBe("sk_test_second");
-    });
-
-    test("getStripeKeyMode returns null when no key is set", () => {
-      expect(settings.stripe.keyMode).toBeNull();
-    });
-
-    test("getStripeKeyMode returns test for sk_test_ key", async () => {
-      await settings.update.stripe.secretKey("sk_test_abc123");
-      expect(settings.stripe.keyMode).toBe("test");
-    });
-
-    test("getStripeKeyMode returns live for sk_live_ key", async () => {
-      await settings.update.stripe.secretKey("sk_live_abc123");
-      expect(settings.stripe.keyMode).toBe("live");
-    });
-
-    test("getStripeKeyMode returns null for unrecognised key prefix", async () => {
-      await settings.update.stripe.secretKey("rk_invalid_abc123");
-      expect(settings.stripe.keyMode).toBeNull();
-    });
-  });
-
-  describe("admin password", () => {
-    test("verifyUserPassword returns hash for correct password", async () => {
-      // Use the user created by createTestDbWithSetup
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      const result = await verifyUserPassword(user!, TEST_ADMIN_PASSWORD);
-      expect(result).toBeTruthy();
-      expect(result).toContain("pbkdf2:");
-    });
-
-    test("verifyUserPassword returns null for wrong password", async () => {
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      const result = await verifyUserPassword(user!, "wrong");
-      expect(result).toBeNull();
-    });
-
-    test("updateUserPassword re-wraps DATA_KEY with new KEK", async () => {
-      // Use the user from createTestDbWithSetup
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      const oldWrappedKey = user!.wrapped_data_key;
-
-      const oldHash = await verifyUserPassword(user!, TEST_ADMIN_PASSWORD);
-      expect(oldHash).toBeTruthy();
-
-      const success = await settings.updateUserPassword(
-        user!.id,
-        oldHash!,
-        user!.wrapped_data_key!,
-        "newpassword456",
-      );
-      expect(success).toBe(true);
-
-      // Wrapped key should be different (re-wrapped with new KEK)
-      const updatedUser = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(updatedUser!.wrapped_data_key).not.toBe(oldWrappedKey);
-
-      // Old password should no longer work
-      expect(
-        await verifyUserPassword(updatedUser!, TEST_ADMIN_PASSWORD),
-      ).toBeNull();
-
-      // New password should work
-      expect(
-        await verifyUserPassword(updatedUser!, "newpassword456"),
-      ).toBeTruthy();
-    });
-
-    test("updateUserPassword fails with wrong old password hash", async () => {
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-
-      const { settings: s } = await import("#lib/db/settings.ts");
-      // Pass a bogus password hash - KEK derivation will produce wrong key
-      const success = await s.updateUserPassword(
-        user!.id,
-        "pbkdf2:bogus:hash",
-        user!.wrapped_data_key!,
-        "newpassword",
-      );
-      expect(success).toBe(false);
-
-      // Original password should still work
-      const unchanged = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(
-        await verifyUserPassword(unchanged!, TEST_ADMIN_PASSWORD),
-      ).toBeTruthy();
-    });
-
-    test("password change allows decryption of both old and new attendee records", async () => {
-      const newPassword = "newpassword456";
-
-      // Use the user from createTestDbWithSetup (TEST_ADMIN_PASSWORD)
-      // Create an event via REST API
-      const event = await createTestEvent({
-        name: "Password Test Event",
-        maxAttendees: 100,
-        thankYouUrl: "https://example.com/thanks",
-      });
-
-      // Create an attendee BEFORE password change
-      const beforeResult = await createAttendeeAtomic({
-        name: "Alice Before",
-        email: "alice@example.com",
-        paymentId: "pi_before_change",
-        bookings: [{ eventId: event.id }],
-      });
-      if (!beforeResult.success) throw new Error("Failed to create attendee");
-      const attendeeBefore = beforeResult.attendees[0]!;
-
-      // Change the password using user-based API
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      const oldHash = await verifyUserPassword(user!, TEST_ADMIN_PASSWORD);
-      expect(oldHash).toBeTruthy();
-
-      const changeSuccess = await settings.updateUserPassword(
-        user!.id,
-        oldHash!,
-        user!.wrapped_data_key!,
-        newPassword,
-      );
-      expect(changeSuccess).toBe(true);
-
-      // Create an attendee AFTER password change
-      const afterResult = await createAttendeeAtomic({
-        name: "Bob After",
-        email: "bob@example.com",
-        paymentId: "pi_after_change",
-        bookings: [{ eventId: event.id }],
-      });
-      if (!afterResult.success) throw new Error("Failed to create attendee");
-      const attendeeAfter = afterResult.attendees[0]!;
-
-      // Get the private key using the NEW password
-      const updatedUser = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(updatedUser).not.toBeNull();
-      const newPasswordHash = await verifyUserPassword(
-        updatedUser!,
-        newPassword,
-      );
-      expect(newPasswordHash).toBeTruthy();
-
-      const kek = await deriveKEK(newPasswordHash!);
-      const dataKey = await unwrapKey(updatedUser!.wrapped_data_key!, kek);
-
-      const wrappedPrivateKey = settings.wrappedPrivateKey;
-      expect(wrappedPrivateKey).toBeTruthy();
-
-      const privateKeyJwk = await decryptWithKey(wrappedPrivateKey!, dataKey);
-      const privateKey = await importPrivateKey(privateKeyJwk);
-
-      // Decrypt the attendee created BEFORE password change
-      const decryptedBefore = await getAttendee(attendeeBefore.id, privateKey);
-      expect(decryptedBefore).not.toBeNull();
-      expect(decryptedBefore?.name).toBe("Alice Before");
-      expect(decryptedBefore?.email).toBe("alice@example.com");
-      expect(decryptedBefore?.payment_id).toBe("pi_before_change");
-
-      // Decrypt the attendee created AFTER password change
-      const decryptedAfter = await getAttendee(attendeeAfter.id, privateKey);
-      expect(decryptedAfter).not.toBeNull();
-      expect(decryptedAfter?.name).toBe("Bob After");
-      expect(decryptedAfter?.email).toBe("bob@example.com");
-      expect(decryptedAfter?.payment_id).toBe("pi_after_change");
-    });
-  });
-
-  describe("events", () => {
-    test("createEvent creates event with correct properties", async () => {
-      const event = await createTestEvent({
-        name: "My Test Event",
-        maxAttendees: 100,
-        thankYouUrl: "https://example.com/thanks",
-      });
-
-      expect(event.id).toBe(1);
-      expect(event.name).toBe("My Test Event");
-      expect(event.slug).toBeDefined();
-      expect(event.max_attendees).toBe(100);
-      expect(event.thank_you_url).toBe("https://example.com/thanks");
-      expect(event.created).toBeDefined();
-      expect(event.unit_price).toBe(0);
-    });
-
-    test("createEvent creates event with unit_price", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com/thanks",
-        unitPrice: 1000,
-      });
-
-      expect(event.unit_price).toBe(1000);
-    });
-
-    test("createEvent stores and retrieves description", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        description: "A test description",
-      });
-
-      expect(event.description).toBe("A test description");
-    });
-
-    test("createEvent defaults description to empty string", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-      });
-
-      expect(event.description).toBe("");
-    });
-
-    test("getAllEvents returns empty array when no events", async () => {
-      const events = await getAllEvents();
-      expect(events).toEqual([]);
-    });
-
-    test("getAllEvents returns events with attendee count", async () => {
-      await createTestEvent({
-        name: "Event One",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestEvent({
-        name: "Event Two",
-        maxAttendees: 100,
-        thankYouUrl: "https://example.com",
-      });
-
-      const events = await getAllEvents();
-      expect(events.length).toBe(2);
-      expect(events[0]?.attendee_count).toBe(0);
-      expect(events[1]?.attendee_count).toBe(0);
-    });
-
-    test("getEvent returns null for missing event", async () => {
-      const event = await getEvent(999);
-      expect(event).toBeNull();
-    });
-
-    test("getEvent returns event by id", async () => {
-      const created = await createTestEvent({
-        name: "Fetch Test",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const fetched = await getEvent(created.id);
-
-      expect(fetched).not.toBeNull();
-      expect(fetched?.name).toBe("Fetch Test");
-    });
-
-    test("getEventWithCount returns null for missing event", async () => {
-      const event = await getEventWithCount(999);
-      expect(event).toBeNull();
-    });
-
-    test("getEventWithCount returns event with count", async () => {
-      const created = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const fetched = await getEventWithCount(created.id);
-
-      expect(fetched).not.toBeNull();
-      expect(fetched?.attendee_count).toBe(0);
-    });
-
-    test("eventsTable.update updates event properties", async () => {
-      const created = await createTestEvent({
-        name: "Original Event",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com/original",
-      });
-
-      const updated = await eventsTable.update(created.id, {
-        name: "Updated Event",
-        slug: created.slug,
-        slugIndex: created.slug_index,
-        maxAttendees: 100,
-        thankYouUrl: "https://example.com/updated",
-        unitPrice: 1500,
-      });
-
-      expect(updated).not.toBeNull();
-      expect(updated?.name).toBe("Updated Event");
-      expect(updated?.max_attendees).toBe(100);
-      expect(updated?.thank_you_url).toBe("https://example.com/updated");
-      expect(updated?.unit_price).toBe(1500);
-    });
-
-    test("eventsTable.update returns null for non-existent event", async () => {
-      const result = await eventsTable.update(999, {
-        name: "Non Existent",
-        slug: "non-existent",
-        slugIndex: "non-existent",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      expect(result).toBeNull();
-    });
-
-    test("eventsTable.update can set unit_price to zero", async () => {
-      const created = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-        unitPrice: 1000,
-      });
-
-      const updated = await eventsTable.update(created.id, {
-        name: created.name,
-        slug: created.slug,
-        slugIndex: created.slug_index,
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-        unitPrice: 0,
-      });
-
-      expect(updated?.unit_price).toBe(0);
-    });
-
-    test("deleteEvent removes event", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      await deleteEvent(event.id);
-
-      const fetched = await getEvent(event.id);
-      expect(fetched).toBeNull();
-    });
-
-    test("deleteEvent removes all attendees for the event", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "John",
-        "john@example.com",
-      );
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "Jane",
-        "jane@example.com",
-      );
-
-      await deleteEvent(event.id);
-
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees).toEqual([]);
-    });
-
-    test("deleteEvent removes processed payment records for attendees", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const attendee = await createTestAttendee(
-        event.id,
-        event.slug,
-        "John Doe",
-        "john@example.com",
-      );
-
-      await reserveSession("sess_event_delete");
-      await finalizePaymentSession("sess_event_delete", attendee.id);
-
-      await deleteEvent(event.id);
-
-      const processed = await isSessionProcessed("sess_event_delete");
-      expect(processed).toBeNull();
-    });
-
-    test("deleteEvent removes activity log entries for the event", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      await logActivity("Action for event", event.id);
-      await logActivity("Another action", event.id);
-      await logActivity("Global action");
-
-      await deleteEvent(event.id);
-
-      const eventLog = await getEventActivityLog(event.id);
-      expect(eventLog).toEqual([]);
-
-      const allLog = await getAllActivityLog();
-      const messages = allLog.map((e) => e.message);
-      expect(messages).not.toContain("Action for event");
-      expect(messages).not.toContain("Another action");
-      expect(messages).toContain("Global action");
-    });
-
-    test("deleteEvent works with no attendees", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      await deleteEvent(event.id);
-
-      const fetched = await getEvent(event.id);
-      expect(fetched).toBeNull();
-    });
-
-    test("deleteEvent preserves attendees linked to other events", async () => {
-      const event1 = await createTestEvent({ maxAttendees: 50 });
-      const event2 = await createTestEvent({ maxAttendees: 50 });
-      // Create one attendee linked to both events
-      const result = await createAttendeeAtomic({
-        name: "Multi",
-        email: "multi@example.com",
-        bookings: [{ eventId: event1.id }, { eventId: event2.id }],
-      });
-      expect(result.success).toBe(true);
-      if (!result.success) return;
-      const attendeeId = result.attendees[0]!.id;
-
-      // Delete event1 — attendee should survive (still linked to event2)
-      await deleteEvent(event1.id);
-
-      const raw = await getAttendeesRaw(event2.id);
-      expect(raw.length).toBe(1);
-      expect(raw[0]!.id).toBe(attendeeId);
-    });
-
-    test("deleteEvent cleans up orphaned attendees", async () => {
-      const event = await createTestEvent({ maxAttendees: 50 });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "Solo",
-        "solo@example.com",
-      );
-
-      await deleteEvent(event.id);
-
-      // Attendee linked only to this event should be deleted
-      const { getDb } = await import("#lib/db/client.ts");
-      const rows = await getDb().execute(
-        "SELECT COUNT(*) as count FROM attendees",
-      );
-      expect(rows.rows[0]!.count).toBe(0);
-    });
-
-    test("unlinkAttendeeFromEvent removes link and preserves attendee", async () => {
-      const { unlinkAttendeeFromEvent } = await import("#lib/db/attendees.ts");
-      const event1 = await createTestEvent({ maxAttendees: 50 });
-      const event2 = await createTestEvent({ maxAttendees: 50 });
-      const result = await createAttendeeAtomic({
-        name: "Unlink",
-        email: "unlink@example.com",
-        bookings: [{ eventId: event1.id }, { eventId: event2.id }],
-      });
-      expect(result.success).toBe(true);
-      if (!result.success) return;
-
-      const { attendeeDeleted } = await unlinkAttendeeFromEvent(
-        result.attendees[0]!.id,
-        event1.id,
-      );
-
-      expect(attendeeDeleted).toBe(false);
-      // Attendee still linked to event2
-      const raw = await getAttendeesRaw(event2.id);
-      expect(raw.length).toBe(1);
-      // No longer linked to event1
-      const raw1 = await getAttendeesRaw(event1.id);
-      expect(raw1.length).toBe(0);
-    });
-
-    test("unlinkAttendeeFromEvent deletes orphaned attendee", async () => {
-      const { unlinkAttendeeFromEvent } = await import("#lib/db/attendees.ts");
-      const event = await createTestEvent({ maxAttendees: 50 });
-      const result = await createAttendeeAtomic({
-        name: "Orphan",
-        email: "orphan@example.com",
-        bookings: [{ eventId: event.id }],
-      });
-      expect(result.success).toBe(true);
-      if (!result.success) return;
-
-      const { attendeeDeleted } = await unlinkAttendeeFromEvent(
-        result.attendees[0]!.id,
-        event.id,
-      );
-
-      expect(attendeeDeleted).toBe(true);
-      const { getDb } = await import("#lib/db/client.ts");
-      const rows = await getDb().execute(
-        "SELECT COUNT(*) as count FROM attendees",
-      );
-      expect(rows.rows[0]!.count).toBe(0);
-    });
-
-    test("eventsTable.deleteById invalidates cache", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Populate cache
-      const before = await getEvent(event.id);
-      expect(before).not.toBeNull();
-
-      await eventsTable.deleteById(event.id);
-
-      const after = await getEvent(event.id);
-      expect(after).toBeNull();
-    });
-
-    test("isSlugTaken with excludeEventId excludes that event", async () => {
-      const event = await createTestEvent({
-        name: "Slug Taken Test",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Slug is taken when checking without exclusion
-      const taken = await isSlugTaken(event.slug);
-      expect(taken).toBe(true);
-
-      // Slug is not taken when excluding the event that owns it
-      const notTaken = await isSlugTaken(event.slug, event.id);
-      expect(notTaken).toBe(false);
-    });
-  });
-
-  describe("attendees", () => {
-    test("getAttendee returns null for missing attendee", async () => {
-      const privateKey = await getTestPrivateKey();
-      const attendee = await getAttendee(999, privateKey);
-      expect(attendee).toBeNull();
-    });
-
-    test("getAttendee returns attendee by id", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const created = await createTestAttendee(
-        event.id,
-        event.slug,
-        "John Doe",
-        "john@example.com",
-      );
-      const privateKey = await getTestPrivateKey();
-      const fetched = await getAttendee(created.id, privateKey);
-
-      expect(fetched).not.toBeNull();
-      expect(fetched?.name).toBe("John Doe");
-    });
-
-    test("deleteAttendee removes attendee", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const attendee = await createTestAttendee(
-        event.id,
-        event.slug,
-        "John Doe",
-        "john@example.com",
-      );
-
-      await deleteAttendee(attendee.id);
-
-      const privateKey = await getTestPrivateKey();
-      const fetched = await getAttendee(attendee.id, privateKey);
-      expect(fetched).toBeNull();
-    });
-
-    test("deleteAttendee removes processed payment records", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const attendee = await createTestAttendee(
-        event.id,
-        event.slug,
-        "Jane Doe",
-        "jane@example.com",
-      );
-
-      await reserveSession("sess_attendee_delete");
-      await finalizePaymentSession("sess_attendee_delete", attendee.id);
-
-      await deleteAttendee(attendee.id);
-
-      const processed = await isSessionProcessed("sess_attendee_delete");
-      expect(processed).toBeNull();
-    });
-
-    test("decryptAttendees returns empty array when no attendees", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees).toEqual([]);
-    });
-
-    test("decryptAttendees returns decrypted attendees for event", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "John",
-        "john@example.com",
-      );
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "Jane",
-        "jane@example.com",
-      );
-
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees.length).toBe(2);
-    });
-
-    test("getNewestAttendeesRaw returns attendees across events ordered by newest first", async () => {
-      const event1 = await createTestEvent({ maxAttendees: 50 });
-      const event2 = await createTestEvent({ maxAttendees: 50 });
-      await createTestAttendee(
-        event1.id,
-        event1.slug,
-        "First",
-        "first@example.com",
-      );
-      await createTestAttendee(
-        event2.id,
-        event2.slug,
-        "Second",
-        "second@example.com",
-      );
-      await createTestAttendee(
-        event1.id,
-        event1.slug,
-        "Third",
-        "third@example.com",
-      );
-
-      const raw = await getNewestAttendeesRaw(10);
-      expect(raw.length).toBe(3);
-      // Newest first (Third created last)
-      expect(raw[0]?.event_id).toBe(event1.id);
-      expect(raw[1]?.event_id).toBe(event2.id);
-      expect(raw[2]?.event_id).toBe(event1.id);
-    });
-
-    test("getNewestAttendeesRaw respects limit", async () => {
-      const event = await createTestEvent({ maxAttendees: 50 });
-      await createTestAttendee(event.id, event.slug, "A", "a@example.com");
-      await createTestAttendee(event.id, event.slug, "B", "b@example.com");
-      await createTestAttendee(event.id, event.slug, "C", "c@example.com");
-
-      const raw = await getNewestAttendeesRaw(2);
-      expect(raw.length).toBe(2);
-    });
-
-    test("getNewestAttendeesRaw returns empty array when no attendees", async () => {
-      const raw = await getNewestAttendeesRaw(10);
-      expect(raw).toEqual([]);
-    });
-
-    test("getActiveEventStats returns zeros for empty events", async () => {
-      const stats = await getActiveEventStats([]);
-      expect(stats).toEqual({ income: 0, tickets: 0, attendees: 0 });
-    });
-
-    test("getActiveEventStats returns zeros when all events inactive", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        unitPrice: 500,
-      });
-      await createPaidTestAttendee(
-        event.id,
-        "Alice",
-        "alice@example.com",
-        "pay_1",
-        1000,
-      );
-      const events = await getAllEvents();
-      const inactive = events.map((e) => ({ ...e, active: false }));
-      const stats = await getActiveEventStats(inactive);
-      expect(stats).toEqual({ income: 0, tickets: 0, attendees: 0 });
-    });
-
-    test("getActiveEventStats counts tickets and sums income for active events", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        unitPrice: 500,
-      });
-      await createPaidTestAttendee(
-        event.id,
-        "Alice",
-        "alice@example.com",
-        "pay_1",
-        1000,
-      );
-      await createPaidTestAttendee(
-        event.id,
-        "Bob",
-        "bob@example.com",
-        "pay_2",
-        2000,
-      );
-      const events = await getAllEvents();
-      const stats = await getActiveEventStats(events);
-      expect(stats.tickets).toBe(2);
-      expect(stats.income).toBe(3000);
-      expect(stats.attendees).toBe(2);
-    });
-
-    test("getActiveEventStats excludes inactive events", async () => {
-      const event1 = await createTestEvent({
-        maxAttendees: 50,
-        unitPrice: 500,
-      });
-      const event2 = await createTestEvent({
-        maxAttendees: 50,
-        unitPrice: 500,
-      });
-      await createPaidTestAttendee(
-        event1.id,
-        "Alice",
-        "alice@example.com",
-        "pay_1",
-        1000,
-      );
-      await createPaidTestAttendee(
-        event2.id,
-        "Bob",
-        "bob@example.com",
-        "pay_2",
-        2000,
-      );
-      const events = await getAllEvents();
-      const mixed = events.map((e) =>
-        e.id === event2.id ? { ...e, active: false } : e,
-      );
-      const stats = await getActiveEventStats(mixed);
-      expect(stats.tickets).toBe(1);
-      expect(stats.income).toBe(1000);
-      expect(stats.attendees).toBe(1);
-    });
-
-    test("getActiveEventStats treats non-numeric price_paid as zero", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        unitPrice: 0,
-      });
-      await createPaidTestAttendee(
-        event.id,
-        "Free Alice",
-        "free@example.com",
-        "",
-        0,
-      );
-      const events = await getAllEvents();
-      const stats = await getActiveEventStats(events);
-      expect(stats.tickets).toBe(1);
-      expect(stats.income).toBe(0);
-      expect(stats.attendees).toBe(1);
-    });
-
-    test("attendee count reflects in getEventWithCount", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "John",
-        "john@example.com",
-      );
-
-      const fetched = await getEventWithCount(event.id);
-      expect(fetched?.attendee_count).toBe(1);
-    });
-
-    test("attendee count reflects in getAllEvents", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "John",
-        "john@example.com",
-      );
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "Jane",
-        "jane@example.com",
-      );
-
-      const events = await getAllEvents();
-      expect(events[0]?.attendee_count).toBe(2);
-    });
-
-    test("createAttendeeAtomic succeeds when capacity available", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 2,
-        thankYouUrl: "https://example.com",
-      });
-
-      const result = await createAttendeeAtomic({
-        name: "John",
-        email: "john@example.com",
-        paymentId: "pi_test",
-        bookings: [{ eventId: event.id, quantity: 1 }],
-      });
-
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.attendees[0]!.name).toBe("John");
-        expect(result.attendees[0]!.payment_id).toBe("pi_test");
-      }
-    });
-
-    test("createAttendeeAtomic links all bookings to correct attendee for multi-event group purchase", async () => {
-      const event1 = await createTestEvent({ maxAttendees: 50 });
-      const event2 = await createTestEvent({ maxAttendees: 50 });
-
-      // Create a prior attendee with TWO bookings so the event_attendees
-      // auto-increment gets ahead of attendees — this prevents the bug from
-      // being masked by coincidental IDs (attendees.id == event_attendees.id)
-      const event3 = await createTestEvent({ maxAttendees: 50 });
-      await createAttendeeAtomic({
-        name: "Prior",
-        email: "prior@example.com",
-        bookings: [{ eventId: event1.id }, { eventId: event3.id }],
-      });
-
-      const result = await createAttendeeAtomic({
-        name: "Group Buyer",
-        email: "group@example.com",
-        paymentId: "pi_group_test",
-        bookings: [
-          { eventId: event1.id, quantity: 1, pricePaid: 1000 },
-          { eventId: event2.id, quantity: 1, pricePaid: 1500 },
-        ],
-      });
-
-      expect(result.success).toBe(true);
-      if (!result.success) return;
-
-      // Both bookings should be reported as successful
-      expect(result.attendees.length).toBe(2);
-      const attendeeId = result.attendees[0]!.id;
-
-      // Verify all event_attendees rows for this attendee point to the correct attendee_id
-      const rows = await getDb().execute({
-        sql: `SELECT attendee_id, event_id
-              FROM event_attendees
-              WHERE attendee_id = ?
-              ORDER BY event_id`,
-        args: [attendeeId],
-      });
-      expect(rows.rows.length).toBe(2);
-
-      // Look up the attendee by token — this is the path the ticket page uses
-      const token = result.attendees[0]!.ticket_token;
-      const lookupResults = await getAttendeesByTokens([token]);
-      const awb = lookupResults[0];
-      expect(awb).toBeDefined();
-
-      // The attendee should have bookings for BOTH events
-      const bookingEventIds = awb!.bookings.map((b) => b.event_id).sort();
-      expect(bookingEventIds).toEqual([event1.id, event2.id].sort());
-    });
-
-    test("createAttendeeAtomic fails when capacity exceeded", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 1,
-        thankYouUrl: "https://example.com",
-      });
-      // Use createAttendeeAtomic to fill capacity (production code path)
-      await createAttendeeAtomic({
-        name: "First",
-        email: "first@example.com",
-        bookings: [{ eventId: event.id }],
-      });
-
-      const result = await createAttendeeAtomic({
-        name: "Second",
-        email: "second@example.com",
-        bookings: [{ eventId: event.id, quantity: 1 }],
-      });
-
-      expect(result.success).toBe(false);
-      if (!result.success) {
-        expect(result.reason).toBe("capacity_exceeded");
-      }
-    });
-
-    test("createAttendeeAtomic fails with empty bookings", async () => {
-      const result = await createAttendeeAtomic({
-        name: "Nobody",
-        email: "nobody@example.com",
-        bookings: [],
-      });
-
-      expect(result.success).toBe(false);
-      if (!result.success) {
-        expect(result.reason).toBe("capacity_exceeded");
-      }
-    });
-
-    test("createAttendeeAtomic fails when encryption key not configured", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Remove public key to simulate incomplete setup
-      await getDb().execute({
-        sql: "DELETE FROM settings WHERE key = ?",
-        args: [CONFIG_KEYS.PUBLIC_KEY],
-      });
-      settings.invalidateCache();
-
-      const result = await createAttendeeAtomic({
-        name: "John",
-        email: "john@example.com",
-        bookings: [{ eventId: event.id }],
-      });
-
-      expect(result.success).toBe(false);
-      if (!result.success) {
-        expect(result.reason).toBe("encryption_error");
-      }
-    });
-  });
-
-  describe("hasAvailableSpots", () => {
-    test("returns false for non-existent event", async () => {
-      const result = await hasAvailableSpots(999);
-      expect(result).toBe(false);
-    });
-
-    test("returns true when spots available", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 2,
-        thankYouUrl: "https://example.com",
-      });
-      const result = await hasAvailableSpots(event.id);
-      expect(result).toBe(true);
-    });
-
-    test("returns true when some spots taken", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 2,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "John",
-        "john@example.com",
-      );
-
-      const result = await hasAvailableSpots(event.id);
-      expect(result).toBe(true);
-    });
-
-    test("returns false when event is full", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 2,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "John",
-        "john@example.com",
-      );
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "Jane",
-        "jane@example.com",
-      );
-
-      const result = await hasAvailableSpots(event.id);
-      expect(result).toBe(false);
-    });
-
-    test("checks per-date capacity for daily events", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 1,
-        eventType: "daily",
-        bookableDays: [
-          "Monday",
-          "Tuesday",
-          "Wednesday",
-          "Thursday",
-          "Friday",
-          "Saturday",
-          "Sunday",
-        ],
-        minimumDaysBefore: 0,
-        maximumDaysAfter: 14,
-      });
-
-      // Create an attendee for a specific date
-      await createAttendeeAtomic({
-        name: "Day User",
-        email: "day@example.com",
-        bookings: [{ eventId: event.id, date: "2026-02-10" }],
-      });
-
-      // That date should be full
-      const full = await hasAvailableSpots(event.id, 1, "2026-02-10");
-      expect(full).toBe(false);
-
-      // A different date should be available
-      const available = await hasAvailableSpots(event.id, 1, "2026-02-11");
-      expect(available).toBe(true);
-    });
-  });
-
-  describe("getDateAttendeeCount", () => {
-    test("returns 0 when no attendees for date", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 10,
-        eventType: "daily",
-        bookableDays: [
-          "Monday",
-          "Tuesday",
-          "Wednesday",
-          "Thursday",
-          "Friday",
-          "Saturday",
-          "Sunday",
-        ],
-        minimumDaysBefore: 0,
-        maximumDaysAfter: 14,
-      });
-      const count = await getDateAttendeeCount(event.id, "2026-02-10");
-      expect(count).toBe(0);
-    });
-
-    test("returns correct count for date with attendees", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 10,
-        eventType: "daily",
-        bookableDays: [
-          "Monday",
-          "Tuesday",
-          "Wednesday",
-          "Thursday",
-          "Friday",
-          "Saturday",
-          "Sunday",
-        ],
-        minimumDaysBefore: 0,
-        maximumDaysAfter: 14,
-      });
-
-      await createAttendeeAtomic({
-        name: "User 1",
-        email: "u1@example.com",
-        bookings: [{ eventId: event.id, quantity: 2, date: "2026-02-10" }],
-      });
-      await createAttendeeAtomic({
-        name: "User 2",
-        email: "u2@example.com",
-        bookings: [{ eventId: event.id, quantity: 3, date: "2026-02-10" }],
-      });
-
-      const count = await getDateAttendeeCount(event.id, "2026-02-10");
-      expect(count).toBe(5);
-    });
-
-    test("does not count attendees on different dates", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 10,
-        eventType: "daily",
-        bookableDays: [
-          "Monday",
-          "Tuesday",
-          "Wednesday",
-          "Thursday",
-          "Friday",
-          "Saturday",
-          "Sunday",
-        ],
-        minimumDaysBefore: 0,
-        maximumDaysAfter: 14,
-      });
-
-      await createAttendeeAtomic({
-        name: "User 1",
-        email: "u1@example.com",
-        bookings: [{ eventId: event.id, quantity: 2, date: "2026-02-10" }],
-      });
-
-      const count = await getDateAttendeeCount(event.id, "2026-02-11");
-      expect(count).toBe(0);
-    });
-  });
-
-  describe("checkBatchAvailability", () => {
-    test("returns true for empty items", async () => {
-      const result = await checkBatchAvailability([], null);
-      expect(result).toBe(true);
-    });
-
-    test("returns false when event not found", async () => {
-      const result = await checkBatchAvailability(
-        [{ eventId: 99999, quantity: 1 }],
-        null,
-      );
-      expect(result).toBe(false);
-    });
-
-    test("checks per-date capacity for daily events", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 2,
-        eventType: "daily",
-        bookableDays: [
-          "Monday",
-          "Tuesday",
-          "Wednesday",
-          "Thursday",
-          "Friday",
-          "Saturday",
-          "Sunday",
-        ],
-        minimumDaysBefore: 0,
-        maximumDaysAfter: 14,
-      });
-
-      await createAttendeeAtomic({
-        name: "Existing",
-        email: "existing@example.com",
-        bookings: [{ eventId: event.id, quantity: 1, date: "2026-02-10" }],
-      });
-
-      // Should have 1 spot left on that date
-      const available = await checkBatchAvailability(
-        [{ eventId: event.id, quantity: 1 }],
-        "2026-02-10",
-      );
-      expect(available).toBe(true);
-
-      // Should not have 2 spots on that date
-      const notAvailable = await checkBatchAvailability(
-        [{ eventId: event.id, quantity: 2 }],
-        "2026-02-10",
-      );
-      expect(notAvailable).toBe(false);
-    });
-  });
-
-  describe("getDb", () => {
-    test("creates client when db is null", () => {
-      setDb(null);
-      const restore = setTestEnv({ DB_URL: ":memory:" });
-      const client = getDb();
-      expect(client).toBeDefined();
-      restore();
-    });
-
-    test("returns existing client when db is set", () => {
-      const client1 = getDb();
-      const client2 = getDb();
-      expect(client1).toBe(client2);
-    });
-  });
-
-  describe("sessions", () => {
-    test("createSession and getSession work together", async () => {
-      const expires = Date.now() + 1000;
-      await createSession("test-token", "test-csrf-token", expires, null, 1);
-
-      const session = await getSession("test-token");
-      expect(session).not.toBeNull();
-      // Token is hashed in storage, verify by csrf_token and expires
-      expect(session?.csrf_token).toBe("test-csrf-token");
-      expect(session?.expires).toBe(expires);
-    });
-
-    test("getSession returns null for missing session", async () => {
-      const session = await getSession("nonexistent");
-      expect(session).toBeNull();
-    });
-
-    test("deleteSession removes session", async () => {
-      await createSession(
-        "delete-me",
-        "csrf-delete",
-        Date.now() + 1000,
-        null,
-        1,
-      );
-      await deleteSession("delete-me");
-
-      const session = await getSession("delete-me");
-      expect(session).toBeNull();
-    });
-
-    test("deleteAllSessions removes all sessions", async () => {
-      await createSession("session1", "csrf1", Date.now() + 10000, null, 1);
-      await createSession("session2", "csrf2", Date.now() + 10000, null, 1);
-      await createSession("session3", "csrf3", Date.now() + 10000, null, 1);
-
-      await deleteAllSessions();
-
-      const session1 = await getSession("session1");
-      const session2 = await getSession("session2");
-      const session3 = await getSession("session3");
-
-      expect(session1).toBeNull();
-      expect(session2).toBeNull();
-      expect(session3).toBeNull();
-    });
-
-    test("getAllSessions returns all sessions ordered by expiration descending", async () => {
-      const now = Date.now();
-      await createSession("session1", "csrf1", now + 1000, null, 1);
-      await createSession("session2", "csrf2", now + 3000, null, 1);
-      await createSession("session3", "csrf3", now + 2000, null, 1);
-
-      const sessions = await getAllSessions();
-
-      expect(sessions.length).toBe(3);
-      // Token is hashed, verify order by csrf_token
-      expect(sessions[0]?.csrf_token).toBe("csrf2"); // Newest first (highest expiry)
-      expect(sessions[1]?.csrf_token).toBe("csrf3");
-      expect(sessions[2]?.csrf_token).toBe("csrf1"); // Oldest last (lowest expiry)
-    });
-
-    test("getAllSessions returns empty array when no sessions", async () => {
-      const sessions = await getAllSessions();
-      expect(sessions).toEqual([]);
-    });
-
-    test("deleteOtherSessions removes all sessions except current", async () => {
-      await createSession(
-        "current",
-        "csrf-current",
-        Date.now() + 10000,
-        null,
-        1,
-      );
-      await createSession("other1", "csrf-other1", Date.now() + 10000, null, 1);
-      await createSession("other2", "csrf-other2", Date.now() + 10000, null, 1);
-
-      await deleteOtherSessions("current");
-
-      const currentSession = await getSession("current");
-      const other1 = await getSession("other1");
-      const other2 = await getSession("other2");
-
-      expect(currentSession).not.toBeNull();
-      expect(other1).toBeNull();
-      expect(other2).toBeNull();
-    });
-
-    test("deleteOtherSessions with no other sessions keeps current", async () => {
-      await createSession("only-session", "csrf", Date.now() + 10000, null, 1);
-
-      await deleteOtherSessions("only-session");
-
-      const session = await getSession("only-session");
-      expect(session).not.toBeNull();
-    });
-
-    test("getSession expires cached entry after TTL", async () => {
-      // Use fake timers to control Date.now()
-      const startTime = Date.now();
-      const time = new FakeTime(startTime);
-
-      try {
-        // Create and cache a session
-        await createSession("ttl-test", "csrf-ttl", startTime + 60000, null, 1);
-        const firstCall = await getSession("ttl-test");
-        expect(firstCall).not.toBeNull();
-
-        // Advance time past the 10-second TTL — cache entry is now stale
-        time.now = startTime + 11000;
-
-        // This call should find the expired cache entry, delete it, and re-query DB
-        const afterTtl = await getSession("ttl-test");
-        expect(afterTtl).not.toBeNull();
-        expect(afterTtl?.csrf_token).toBe("csrf-ttl");
-      } finally {
-        time.restore();
-      }
-    });
-  });
-
-  describe("updateUserPassword", () => {
-    test("updates password and invalidates all sessions", async () => {
-      // Use user from createTestDbWithSetup
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-
-      // Create some sessions
-      await createSession("session1", "csrf1", Date.now() + 10000, null, 1);
-      await createSession("session2", "csrf2", Date.now() + 10000, null, 1);
-
-      // Verify initial password works
-      const initialHash = await verifyUserPassword(user!, TEST_ADMIN_PASSWORD);
-      expect(initialHash).toBeTruthy();
-
-      // Update password using user-based API
-      const success = await settings.updateUserPassword(
-        user!.id,
-        initialHash!,
-        user!.wrapped_data_key!,
-        "new-password-123",
-      );
-      expect(success).toBe(true);
-
-      // Verify new password works
-      const updatedUser = await getUserByUsername(TEST_ADMIN_USERNAME);
-      const newValid = await verifyUserPassword(
-        updatedUser!,
-        "new-password-123",
-      );
-      expect(newValid).toBeTruthy();
-
-      // Verify old password no longer works
-      const oldValid = await verifyUserPassword(
-        updatedUser!,
-        TEST_ADMIN_PASSWORD,
-      );
-      expect(oldValid).toBeNull();
-
-      // Verify all sessions were invalidated
-      const session1 = await getSession("session1");
-      const session2 = await getSession("session2");
-      expect(session1).toBeNull();
-      expect(session2).toBeNull();
-    });
-  });
-
-  describe("rate limiting", () => {
-    test("isLoginRateLimited returns false for new IP", async () => {
-      const limited = await isLoginRateLimited("192.168.1.1");
-      expect(limited).toBe(false);
-    });
-
-    test("recordFailedLogin increments attempts", async () => {
-      const locked1 = await recordFailedLogin("192.168.1.2");
-      expect(locked1).toBe(false);
-
-      const locked2 = await recordFailedLogin("192.168.1.2");
-      expect(locked2).toBe(false);
-    });
-
-    test("recordFailedLogin locks after 5 attempts", async () => {
-      for (let i = 0; i < 4; i++) {
-        const locked = await recordFailedLogin("192.168.1.3");
-        expect(locked).toBe(false);
-      }
-
-      // 5th attempt should lock
-      const locked = await recordFailedLogin("192.168.1.3");
-      expect(locked).toBe(true);
-    });
-
-    test("isLoginRateLimited returns true when locked", async () => {
-      // Lock the IP
-      for (let i = 0; i < 5; i++) {
-        await recordFailedLogin("192.168.1.4");
-      }
-
-      const limited = await isLoginRateLimited("192.168.1.4");
-      expect(limited).toBe(true);
-    });
-
-    test("clearLoginAttempts clears attempts", async () => {
-      await recordFailedLogin("192.168.1.5");
-      await recordFailedLogin("192.168.1.5");
-
-      await clearLoginAttempts("192.168.1.5");
-
-      // After clearing, should not be limited
-      const limited = await isLoginRateLimited("192.168.1.5");
-      expect(limited).toBe(false);
-    });
-
-    test("isLoginRateLimited clears expired lockout", async () => {
-      // Insert a record with expired lockout
-      await getDb().execute(
-        insert("login_attempts", {
-          ip: "192.168.1.6",
-          attempts: 5,
-          locked_until: Date.now() - 1000,
-        }),
-      );
-
-      // Should clear the expired lockout and return false
-      const limited = await isLoginRateLimited("192.168.1.6");
-      expect(limited).toBe(false);
-    });
-
-    test("isLoginRateLimited returns false for attempts below max without lockout", async () => {
-      // Insert a record with some attempts but no lockout
-      await getDb().execute(
-        insert("login_attempts", {
-          ip: "192.168.1.7",
-          attempts: 3,
-          locked_until: null,
-        }),
-      );
-
-      const limited = await isLoginRateLimited("192.168.1.7");
-      expect(limited).toBe(false);
-    });
-  });
-
-  describe("table utilities", () => {
-    test("toCamelCase converts snake_case to camelCase", async () => {
-      const { toCamelCase } = await import("#lib/db/table.ts");
-      expect(toCamelCase("max_attendees")).toBe("maxAttendees");
-      expect(toCamelCase("thank_you_url")).toBe("thankYouUrl");
-      expect(toCamelCase("name")).toBe("name");
-      expect(toCamelCase("payment_id")).toBe("paymentId");
-    });
-
-    test("toSnakeCase converts camelCase to snake_case", async () => {
-      const { toSnakeCase } = await import("#lib/db/table.ts");
-      expect(toSnakeCase("maxAttendees")).toBe("max_attendees");
-      expect(toSnakeCase("thankYouUrl")).toBe("thank_you_url");
-      expect(toSnakeCase("name")).toBe("name");
-      expect(toSnakeCase("paymentId")).toBe("payment_id");
-    });
-
-    test("buildInputKeyMap creates mapping from DB columns to input keys", async () => {
-      const { buildInputKeyMap } = await import("#lib/db/table.ts");
-      const columns = ["max_attendees", "thank_you_url", "name"];
-      const map = buildInputKeyMap(columns);
-      expect(map).toEqual({
-        max_attendees: "maxAttendees",
-        thank_you_url: "thankYouUrl",
-        name: "name",
-      });
-    });
-
-    test("col.generated creates generated column definition", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const def = col.generated<number>();
-      expect(def.generated).toBe(true);
-    });
-
-    test("col.withDefault creates column with default", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const def = col.withDefault(() => "default-value");
-      expect(def.default?.()).toBe("default-value");
-    });
-
-    test("col.simple creates empty column definition", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const def = col.simple<string>();
-      expect(def).toEqual({});
-    });
-
-    test("col.transform creates column with custom transforms", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const write = (v: string) => v.toUpperCase();
-      const read = (v: string) => v.toLowerCase();
-      const def = col.transform(write, read);
-      expect(def.write?.("hello")).toBe("HELLO");
-      expect(def.read?.("HELLO")).toBe("hello");
-    });
-
-    test("col.encrypted creates column with encrypt/decrypt transforms", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const encrypt = (v: string) => Promise.resolve(`enc:${v}`);
-      const decrypt = (v: string) => Promise.resolve(v.replace("enc:", ""));
-      const def = col.encrypted(encrypt, decrypt);
-      expect(await def.write?.("hello")).toBe("enc:hello");
-      expect(await def.read?.("enc:hello")).toBe("hello");
-    });
-
-    test("col.encryptedText passes through empty strings without encrypting", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const encrypt = (v: string) => Promise.resolve(`enc:${v}`);
-      const decrypt = (v: string) => Promise.resolve(v.replace("enc:", ""));
-      const def = col.encryptedText(encrypt, decrypt);
-      expect(def.default?.()).toBe("");
-      expect(await def.write?.("")).toBe("");
-      expect(await def.read?.("")).toBe("");
-      expect(await def.write?.("hello")).toBe("enc:hello");
-      expect(await def.read?.("enc:hello")).toBe("hello");
-    });
-
-    test("col.encryptedNullable wrapping simple column has no transforms", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const def = col.encryptedNullable(col.simple());
-      expect(def.write).toBeUndefined();
-      expect(def.read).toBeUndefined();
-    });
-
-    test("col.encryptedNullable handles null values", async () => {
-      const { col } = await import("#lib/db/table.ts");
-      const encrypt = (v: string) => Promise.resolve(`enc:${v}`);
-      const decrypt = (v: string) => Promise.resolve(v.replace("enc:", ""));
-      const def = col.encryptedNullable(col.encrypted(encrypt, decrypt));
-      expect(await def.write?.(null)).toBe(null);
-      expect(await def.read?.(null)).toBe(null);
-      expect(await def.write?.("hello")).toBe("enc:hello");
-      expect(await def.read?.("enc:hello")).toBe("hello");
-    });
-
-    test("defineTable.findAll returns all rows", async () => {
-      const { col, defineTable } = await import("#lib/db/table.ts");
-
-      // Create a simple test table
-      type TestRow = { id: number; name: string };
-      type TestInput = { name: string };
-      const testTable = defineTable<TestRow, TestInput>({
-        name: "events",
-        primaryKey: "id",
-        schema: {
-          id: col.generated<number>(),
-          name: col.simple<string>(),
-        },
-      });
-
-      // Create some test events directly
-      await createTestEvent({
-        name: "Event One",
-        maxAttendees: 10,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestEvent({
-        name: "Event Two",
-        maxAttendees: 20,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Use findAll to get all rows
-      const rows = await testTable.findAll();
-      expect(rows.length).toBe(2);
-    });
-
-    test("defineTable.update with no changes returns existing row", async () => {
-      const { col, defineTable } = await import("#lib/db/table.ts");
-
-      // Create an event first
-      const event = await createTestEvent({
-        maxAttendees: 10,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Create a table with only optional fields in input
-      type TestRow = { id: number; name: string };
-      type TestInput = { name?: string };
-      const testTable = defineTable<TestRow, TestInput>({
-        name: "events",
-        primaryKey: "id",
-        schema: {
-          id: col.generated<number>(),
-          name: col.simple<string>(),
-        },
-      });
-
-      // Update with empty input should return the existing row
-      const result = await testTable.update(event.id, {});
-      expect(result).not.toBeNull();
-      expect(result?.id).toBe(event.id);
-    });
-
-    test("defineTable with write transform transforms values on insert", async () => {
-      const { col, defineTable } = await import("#lib/db/table.ts");
-      // Create a table with a write transform that uppercases slug
-      type TestRow = {
-        id: number;
-        slug: string;
-        slug_index: string;
-        created: string;
-        max_attendees: number;
-        thank_you_url: string;
-        unit_price: number;
-        max_quantity: number;
-        webhook_url: string | null;
-        active: number;
-      };
-      type TestInput = {
-        slug: string;
-        slugIndex: string;
-        maxAttendees: number;
-        thankYouUrl: string;
-        unitPrice?: number;
-        maxQuantity?: number;
-        webhookUrl?: string | null;
-        active?: number;
-      };
-      const testTable = defineTable<TestRow, TestInput>({
-        name: "events",
-        primaryKey: "id",
-        schema: {
-          id: col.generated<number>(),
-          created: col.withDefault(() => new Date().toISOString()),
-          slug: col.transform(
-            (v: string) => v.toUpperCase(),
-            (v: string) => v.toLowerCase(),
-          ),
-          slug_index: col.simple<string>(),
-          max_attendees: col.simple<number>(),
-          thank_you_url: col.simple<string>(),
-          unit_price: col.withDefault(() => 0),
-          max_quantity: col.withDefault(() => 1),
-          webhook_url: col.simple<string | null>(),
-          active: col.withDefault(() => 1),
-        },
-      });
-
-      // Insert should apply the write transform
-      const row = await testTable.insert({
-        slug: "test-event",
-        slugIndex: "test-index",
-        maxAttendees: 10,
-        thankYouUrl: "http://test.com",
-      });
-      expect(row.slug).toBe("test-event"); // Returns original input value
-
-      // But the DB should have the transformed value (read transform lowercases)
-      const fromDb = await testTable.findById(row.id);
-      expect(fromDb?.slug).toBe("test-event"); // Read transform lowercases the uppercased "TEST-EVENT"
-    });
-  });
-
-  describe("activity log", () => {
-    test("logActivity creates log entry with message", async () => {
-      const entry = await logActivity("Test action");
-
-      expect(entry.id).toBe(1);
-      expect(entry.message).toBe("Test action");
-      expect(entry.event_id).toBeNull();
-      expect(entry.created).toBeDefined();
-    });
-
-    test("logActivity creates log entry with event ID", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const entry = await logActivity("Created event 'Test Event'", event.id);
-
-      expect(entry.event_id).toBe(event.id);
-      expect(entry.message).toBe("Created event 'Test Event'");
-    });
-
-    test("getEventActivityLog returns entries for specific event", async () => {
-      const event1 = await createTestEvent({
-        name: "Event One",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const event2 = await createTestEvent({
-        name: "Event Two",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      await logActivity("Action for event 1", event1.id);
-      await logActivity("Another action for event 1", event1.id);
-      await logActivity("Action for event 2", event2.id);
-
-      const event1Log = await getEventActivityLog(event1.id);
-      // REST API also logs event creation, so we have 3 entries for event 1
-      expect(event1Log.length).toBe(3);
-      expect(event1Log[0]?.message).toBe("Another action for event 1");
-      expect(event1Log[1]?.message).toBe("Action for event 1");
-    });
-
-    test("getEventActivityLog returns empty array when no entries", async () => {
-      const entries = await getEventActivityLog(999);
-      expect(entries).toEqual([]);
-    });
-
-    test("getEventActivityLog respects limit", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      await logActivity("Action 1", event.id);
-      await logActivity("Action 2", event.id);
-      await logActivity("Action 3", event.id);
-
-      const entries = await getEventActivityLog(event.id, 2);
-      expect(entries.length).toBe(2);
-    });
-
-    test("getAllActivityLog returns all entries", async () => {
-      const event = await createTestEvent({
-        name: "Test Event",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      await logActivity("Global action");
-      await logActivity("Event action", event.id);
-
-      const entries = await getAllActivityLog();
-      // REST API logs event creation, so we have 3 entries total
-      expect(entries.length).toBe(3);
-    });
-
-    test("getAllActivityLog returns entries in descending order", async () => {
-      await logActivity("First action");
-      await logActivity("Second action");
-      await logActivity("Third action");
-
-      const entries = await getAllActivityLog();
-      expect(entries[0]?.message).toBe("Third action");
-      expect(entries[1]?.message).toBe("Second action");
-      expect(entries[2]?.message).toBe("First action");
-    });
-
-    test("getAllActivityLog respects limit", async () => {
-      await logActivity("Action 1");
-      await logActivity("Action 2");
-      await logActivity("Action 3");
-
-      const entries = await getAllActivityLog(2);
-      expect(entries.length).toBe(2);
-    });
-
-    test("getEventWithActivityLog returns event and activity log together", async () => {
-      const event = await createTestEvent({
-        name: "Batch Test Event",
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      await logActivity("First action", event.id);
-      await logActivity("Second action", event.id);
-
-      const result = await getEventWithActivityLog(event.id);
-      expect(result).not.toBeNull();
-      expect(result?.event.id).toBe(event.id);
-      expect(result?.event.name).toBe("Batch Test Event");
-      expect(result?.event.attendee_count).toBe(0);
-      // REST API logs event creation + our 2 = 3
-      expect(result?.entries.length).toBe(3);
-      expect(result?.entries[0]?.message).toBe("Second action");
-      expect(result?.entries[1]?.message).toBe("First action");
-    });
-
-    test("getEventWithActivityLog returns null for non-existent event", async () => {
-      const result = await getEventWithActivityLog(999);
-      expect(result).toBeNull();
-    });
-  });
-
-  describe("attendees - phone decryption", () => {
-    test("decryptAttendees decrypts phone when present", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Create attendee with a non-empty phone to exercise the phone decryption branch
-      const result = await createAttendeeAtomic({
-        name: "Phone Person",
-        email: "phone@example.com",
-        phone: "+44 7700 900000",
-        bookings: [{ eventId: event.id, quantity: 1 }],
-      });
-
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.attendees[0]!.phone).toBe("+44 7700 900000");
-      }
-
-      // Decrypt and verify phone is correctly decrypted
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees.length).toBe(1);
-      expect(attendees[0]?.phone).toBe("+44 7700 900000");
-      expect(attendees[0]?.name).toBe("Phone Person");
-    });
-  });
-
-  describe("attendees - edge cases", () => {
-    test("decryptAttendees handles empty email, phone, address, and special_instructions strings", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Create attendee with empty email/phone/address/special_instructions via createAttendeeAtomic
-      const result = await createAttendeeAtomic({
-        name: "NoContact Person",
-        email: "", // empty email
-        phone: "", // empty phone
-        address: "", // empty address
-        special_instructions: "", // empty special instructions
-        bookings: [{ eventId: event.id, quantity: 1 }],
-      });
-
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.attendees[0]!.email).toBe("");
-        expect(result.attendees[0]!.phone).toBe("");
-        expect(result.attendees[0]!.address).toBe("");
-        expect(result.attendees[0]!.special_instructions).toBe("");
-      }
-
-      // Decrypt and verify empty strings come back correctly
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees.length).toBe(1);
-      expect(attendees[0]?.email).toBe("");
-      expect(attendees[0]?.phone).toBe("");
-      expect(attendees[0]?.address).toBe("");
-      expect(attendees[0]?.special_instructions).toBe("");
-    });
-
-    test("encrypts and decrypts non-empty special_instructions", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      const result = await createAttendeeAtomic({
-        name: "Instructions Person",
-        email: "inst@example.com",
-        special_instructions: "No nuts please\nAllergic to dairy",
-        bookings: [{ eventId: event.id, quantity: 1 }],
-      });
-
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.attendees[0]!.special_instructions).toBe(
-          "No nuts please\nAllergic to dairy",
-        );
-      }
-
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees.length).toBe(1);
-      expect(attendees[0]?.special_instructions).toBe(
-        "No nuts please\nAllergic to dairy",
-      );
-    });
-
-    test("encrypts and decrypts non-empty address", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-
-      const result = await createAttendeeAtomic({
-        name: "Address Person",
-        email: "addr@example.com",
-        address: "123 Main St\nSpringfield\nIL 62701",
-        bookings: [{ eventId: event.id, quantity: 1 }],
-      });
-
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.attendees[0]!.address).toBe(
-          "123 Main St\nSpringfield\nIL 62701",
-        );
-      }
-
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees.length).toBe(1);
-      expect(attendees[0]?.address).toBe("123 Main St\nSpringfield\nIL 62701");
-    });
-
-    test("createAttendeeAtomic stores and returns price_paid when provided", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-        unitPrice: 2500,
-      });
-
-      const result = await createAttendeeAtomic({
-        name: "Paying Customer",
-        email: "pay@example.com",
-        paymentId: "pi_test_price",
-        bookings: [{ eventId: event.id, quantity: 1, pricePaid: 2500 }],
-      });
-
-      expect(result.success).toBe(true);
-      if (result.success) {
-        expect(result.attendees[0]!.price_paid).toBe("2500");
-      }
-
-      // Verify decrypted price_paid
-      const privateKey = await getTestPrivateKey();
-      const raw = await getAttendeesRaw(event.id);
-      const attendees = await decryptAttendees(raw, privateKey);
-      expect(attendees[0]?.price_paid).toBe("2500");
-    });
-  });
-
-  describe("events - batch queries", () => {
-    test("getEventWithAttendeesRaw returns event with attendees", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "Alice",
-        "alice@example.com",
-      );
-
-      const result = await getEventWithAttendeesRaw(event.id);
-      expect(result).not.toBeNull();
-      expect(result?.event.id).toBe(event.id);
-      expect(result?.event.attendee_count).toBe(1);
-      expect(result?.attendeesRaw.length).toBe(1);
-    });
-
-    test("getEventWithAttendeesRaw returns null for non-existent event", async () => {
-      const result = await getEventWithAttendeesRaw(999);
-      expect(result).toBeNull();
-    });
-
-    test("getEventWithAttendeeRaw returns event with count fallback", async () => {
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const attendee = await createTestAttendee(
-        event.id,
-        event.slug,
-        "Bob",
-        "bob@example.com",
-      );
-
-      const result = await getEventWithAttendeeRaw(event.id, attendee.id);
-      expect(result).not.toBeNull();
-      expect(result?.event.id).toBe(event.id);
-      expect(result?.attendeeRaw).not.toBeNull();
-      expect(result?.event.attendee_count).toBe(1);
-    });
-
-    test("getEventWithAttendeeRaw returns null for non-existent event", async () => {
-      const result = await getEventWithAttendeeRaw(999, 1);
-      expect(result).toBeNull();
-    });
-
-    test("getEventsBySlugsBatch returns empty array for empty slugs", async () => {
-      const result = await getEventsBySlugsBatch([]);
-      expect(result).toEqual([]);
-    });
-
-    test("getEventsBySlugsBatch returns events in slug order", async () => {
-      const event1 = await createTestEvent({
-        name: "Batch A",
-        maxAttendees: 10,
-        thankYouUrl: "https://example.com",
-      });
-      const event2 = await createTestEvent({
-        name: "Batch B",
-        maxAttendees: 20,
-        thankYouUrl: "https://example.com",
-      });
-
-      const results = await getEventsBySlugsBatch([event2.slug, event1.slug]);
-      expect(results.length).toBe(2);
-      expect(results[0]?.id).toBe(event2.id);
-      expect(results[1]?.id).toBe(event1.id);
-    });
-
-    test("getEventsBySlugsBatch returns null for missing slugs", async () => {
-      const event = await createTestEvent({
-        name: "Exists",
-        maxAttendees: 10,
-        thankYouUrl: "https://example.com",
-      });
-
-      const results = await getEventsBySlugsBatch([event.slug, "missing"]);
-      expect(results.length).toBe(2);
-      expect(results[0]).not.toBeNull();
-      expect(results[1]).toBeNull();
-    });
-
-    test("getAttendeesByTokens returns attendees in token order", async () => {
-      const event = await createTestEvent({ maxAttendees: 10 });
-
-      // Create attendees directly and get plaintext tokens (matches production flow)
-      const { createTestAttendeeDirect } = await import("#test-utils");
-      const { attendee: a1, token: token1 } = await createTestAttendeeDirect(
-        event.id,
-        "Tok1",
-        "tok1@example.com",
-      );
-      const { attendee: a2, token: token2 } = await createTestAttendeeDirect(
-        event.id,
-        "Tok2",
-        "tok2@example.com",
-      );
-
-      const results = await getAttendeesByTokens([token2, token1]);
-      expect(results.length).toBe(2);
-      expect(results[0]?.id).toBe(a2.id);
-      expect(results[1]?.id).toBe(a1.id);
-    });
-
-    test("getAttendeesByTokens returns null for missing tokens", async () => {
-      const results = await getAttendeesByTokens(["nonexistent"]);
-      expect(results.length).toBe(1);
-      expect(results[0]).toBeNull();
-    });
-
-    test("getAttendeesByTokens returns empty bookings for orphaned attendee", async () => {
-      const event = await createTestEvent({ maxAttendees: 10 });
-      const { createTestAttendeeDirect: createDirect } = await import(
-        "#test-utils"
-      );
-      const { attendee, token } = await createDirect(
-        event.id,
-        "Orphan",
-        "orphan@test.com",
-      );
-      // Remove event link to create orphan
-      await getDb().execute({
-        sql: "DELETE FROM event_attendees WHERE attendee_id = ?",
-        args: [attendee.id],
-      });
-      const results = await getAttendeesByTokens([token]);
-      expect(results[0]).not.toBeNull();
-      expect(results[0]!.bookings).toEqual([]);
-    });
-  });
-
-  describe("login-attempts - expired lockout", () => {
-    test("isLoginRateLimited resets expired lockout and returns false", async () => {
-      // Insert a record with locked_until in the past
-      await getDb().execute(
-        insert("login_attempts", {
-          ip: "expired-ip-hash",
-          attempts: 5,
-          locked_until: Date.now() - 60000,
-        }),
-      );
-
-      // This uses the raw hashed IP - we need to test via the public API
-      // The existing test at line 993 already covers this via isLoginRateLimited
-      // But let's verify that after the expired lockout reset, new attempts work
-      const ip = "192.168.99.1";
-
-      // Lock the IP
-      for (let i = 0; i < 5; i++) {
-        await recordFailedLogin(ip);
-      }
-      expect(await isLoginRateLimited(ip)).toBe(true);
-
-      // Simulate expired lockout by manipulating the DB directly
-      await getDb().execute({
-        sql: `UPDATE login_attempts
-              SET locked_until = ?
-              WHERE locked_until IS NOT NULL`,
-        args: [Date.now() - 1000],
-      });
-
-      // Should detect expired lockout, clear it, and return false
-      const limited = await isLoginRateLimited(ip);
-      expect(limited).toBe(false);
-
-      // Verify the record was cleared - can fail new attempts again
-      const locked = await recordFailedLogin(ip);
-      expect(locked).toBe(false);
-    });
-  });
-
-  describe("processed payments", () => {
-    test("reserveSession succeeds on first call", async () => {
-      const result = await reserveSession("sess_test_1");
-      expect(result.reserved).toBe(true);
-    });
-
-    test("reserveSession returns existing when session already reserved and finalized", async () => {
-      // Create an actual attendee so FK constraint is satisfied
-      const event = await createTestEvent({
-        maxAttendees: 50,
-        thankYouUrl: "https://example.com",
-      });
-      const attendeeResult = await createAttendeeAtomic({
-        name: "Test",
-        email: "test@example.com",
-        bookings: [{ eventId: event.id }],
-      });
-      if (!attendeeResult.success) throw new Error("Failed to create attendee");
-
-      await reserveSession("sess_dup");
-      await finalizePaymentSession("sess_dup", attendeeResult.attendees[0]!.id);
-
-      const result = await reserveSession("sess_dup");
-      expect(result.reserved).toBe(false);
-      if (!result.reserved) {
-        expect(result.existing.attendee_id).toBe(
-          attendeeResult.attendees[0]!.id,
-        );
-      }
-    });
-
-    test("reserveSession handles UNIQUE constraint with unfinalized reservation (not stale)", async () => {
-      // Reserve but don't finalize (attendee_id is NULL)
-      await reserveSession("sess_unfinalized");
-
-      // Immediately try to reserve again - should return existing (not stale yet)
-      const result = await reserveSession("sess_unfinalized");
-      expect(result.reserved).toBe(false);
-      if (!result.reserved) {
-        expect(result.existing.attendee_id).toBeNull();
-      }
-    });
-
-    test("reserveSession retries when stale reservation detected", async () => {
-      // Insert a reservation with a timestamp old enough relative to frozen nowMs
-      const oldTimestamp = new Date(
-        nowMs() - STALE_RESERVATION_MS - 1000,
-      ).toISOString();
-      await getDb().execute(
-        insert("processed_payments", {
-          payment_session_id: "sess_stale",
-          attendee_id: null,
-          processed_at: oldTimestamp,
-        }),
-      );
-
-      // This should detect the stale reservation, delete it, and retry
-      const result = await reserveSession("sess_stale");
-      expect(result.reserved).toBe(true);
-    });
-
-    test("reserveSession re-throws non-unique-constraint errors", async () => {
-      // Drop the table to cause a different error
-      await getDb().execute("DROP TABLE processed_payments");
-
-      try {
-        await reserveSession("sess_error");
-        throw new Error("should not reach here");
-      } catch (e) {
-        expect(String(e)).not.toContain("should not reach here");
-        expect(String(e)).not.toContain("UNIQUE constraint");
-      }
-
-      // Recreate the table for subsequent tests (without FK for simplicity)
-      await getDb().execute(`
-        CREATE TABLE IF NOT EXISTS processed_payments (
-          payment_session_id TEXT PRIMARY KEY,
-          attendee_id INTEGER,
-          processed_at TEXT NOT NULL,
-          ticket_tokens TEXT NOT NULL DEFAULT '',
-          FOREIGN KEY (attendee_id) REFERENCES attendees(id)
-        )
-      `);
-    });
-
-    test("reserveSession retries when stale reservation is detected (recursive path)", async () => {
-      // Insert a reservation with a timestamp old enough relative to frozen nowMs
-      const oldTimestamp = new Date(
-        nowMs() - STALE_RESERVATION_MS - 1000,
-      ).toISOString();
-      await getDb().execute(
-        insert("processed_payments", {
-          payment_session_id: "sess_race",
-          attendee_id: null,
-          processed_at: oldTimestamp,
-        }),
-      );
-
-      const result = await reserveSession("sess_race");
-      expect(result.reserved).toBe(true);
-
-      // Verify the session was re-reserved
-      const processed = await isSessionProcessed("sess_race");
-      expect(processed).not.toBeNull();
-    });
-  });
-
-  describe("settings - additional coverage", () => {
-    test("clearPaymentProvider removes payment provider setting", async () => {
-      await settings.update.paymentProvider("stripe");
-      await settings.loadAll();
-      expect(settings.getCachedRaw(CONFIG_KEYS.PAYMENT_PROVIDER)).toBe(
-        "stripe",
-      );
-
-      await settings.update.clearPaymentProvider();
-      await settings.loadAll();
-      expect(settings.getCachedRaw(CONFIG_KEYS.PAYMENT_PROVIDER)).toBeNull();
-    });
-
-    test("loadAll sets theme to dark when stored value is dark", async () => {
-      await settings.setRaw(CONFIG_KEYS.THEME, "dark");
-      settings.invalidateCache();
-      await settings.loadAll();
-      expect(settings.theme).toBe("dark");
-    });
-
-    test("update.bookingFee with empty string resets to 0", async () => {
-      await settings.update.bookingFee("500");
-      expect(settings.bookingFee).toBe("500");
-      await settings.update.bookingFee("");
-      expect(settings.bookingFee).toBe("0");
-    });
-
-    test("update.stripe.secretKey with empty string sets empty string", async () => {
-      await settings.update.stripe.secretKey("sk_test_abc");
-      expect(settings.stripe.secretKey).toBe("sk_test_abc");
-      await settings.update.stripe.secretKey("");
-      expect(settings.stripe.secretKey).toBe("");
-    });
-
-    test("update.square.accessToken with empty string sets empty string", async () => {
-      await settings.update.square.accessToken("token_123");
-      expect(settings.square.accessToken).toBe("token_123");
-      await settings.update.square.accessToken("");
-      expect(settings.square.accessToken).toBe("");
-    });
-
-    test("update.square.webhookSignatureKey with empty string sets empty string", async () => {
-      await settings.update.square.webhookSignatureKey("sig_key_123");
-      expect(settings.square.webhookSignatureKey).toBe("sig_key_123");
-      await settings.update.square.webhookSignatureKey("");
-      expect(settings.square.webhookSignatureKey).toBe("");
-    });
-
-    test("update.square.locationId with empty string sets empty string", async () => {
-      await settings.update.square.locationId("loc_123");
-      expect(settings.square.locationId).toBe("loc_123");
-      await settings.update.square.locationId("");
-      expect(settings.square.locationId).toBe("");
-    });
-
-    test("updateUserPassword returns false when dataKey unwrap fails", async () => {
-      // Use the user from createTestDbWithSetup
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      const passwordHash = await verifyUserPassword(user!, TEST_ADMIN_PASSWORD);
-      expect(passwordHash).toBeTruthy();
-
-      // Pass corrupted wrapped_data_key - unwrap will fail
-      const { settings: s } = await import("#lib/db/settings.ts");
-      const result = await s.updateUserPassword(
-        user!.id,
-        passwordHash!,
-        "corrupted_wrapped_data_key",
-        "newpassword",
-      );
-      expect(result).toBe(false);
-    });
-  });
-
-  describe("timezone cache", () => {
-    beforeEach(() => {
-      settings.clearTestOverrides();
-    });
-
-    test("getTimezoneCached returns default when no cache exists", () => {
-      settings.invalidateCache();
-      expect(settings.timezone).toBe("Europe/London");
-    });
-
-    test("getTimezoneFromDb returns default when no country is stored", async () => {
-      // Remove the country setting
-      await getDb().execute({
-        sql: "DELETE FROM settings WHERE key = ?",
-        args: [CONFIG_KEYS.COUNTRY],
-      });
-      settings.invalidateCache();
-      const value = settings.timezone;
-      expect(value).toBe("Europe/London");
-    });
-
-    test("getTimezoneCached reads default from TTL cache when no country is stored", async () => {
-      // Remove the country setting
-      await getDb().execute({
-        sql: "DELETE FROM settings WHERE key = ?",
-        args: [CONFIG_KEYS.COUNTRY],
-      });
-      settings.invalidateCache();
-      // Load the settings cache (without country key)
-      settings.getCachedRaw(CONFIG_KEYS.COUNTRY);
-      // Permanent cache should not be set, so it reads from TTL cache
-      // TTL cache has no country key → falls back to default
-      expect(settings.timezone).toBe("Europe/London");
-    });
-
-    test("getTimezoneCached returns value after getTimezoneFromDb populates cache", async () => {
-      await settings.update.country("US");
-      settings.invalidateCache();
-      await settings.loadAll();
-      const value = settings.timezone;
-      expect(value).toBe("America/New_York");
-      expect(settings.timezone).toBe("America/New_York");
-    });
-
-    test("getTimezoneCached reads from TTL cache when permanent cache is empty", async () => {
-      await settings.update.country("JP");
-      settings.invalidateCache();
-      await settings.loadAll();
-      // Force TTL cache to load by calling getSetting
-      settings.getCachedRaw(CONFIG_KEYS.COUNTRY);
-      expect(settings.timezone).toBe("Asia/Tokyo");
-    });
-
-    test("updateCountry updates the permanent cache immediately", async () => {
-      await settings.update.country("NZ");
-      expect(settings.timezone).toBe("Pacific/Auckland");
-    });
-
-    test("getTimezoneFromDb returns test override when set", () => {
-      settings.setForTest({ timezone: "America/Chicago" });
-      const value = settings.timezone;
-      expect(value).toBe("America/Chicago");
-    });
-
-    test("getTimezoneFromDb returns permanent cache when set", () => {
-      // Populate permanent cache via getTimezoneFromDb
-      const value = settings.timezone;
-      // Now the permanent cache is set; calling again should return from cache
-      const cached = settings.timezone;
-      expect(cached).toBe(value);
-    });
-  });
-
-  describe("table utilities - non-generated primary key", () => {
-    test("insert with non-generated primary key uses empty initial row", async () => {
-      const { col, defineTable } = await import("#lib/db/table.ts");
-
-      // Create a table where the primary key is NOT generated (user-supplied)
-      await getDb().execute(`
-        CREATE TABLE IF NOT EXISTS kv_store (
-          key TEXT PRIMARY KEY,
-          value TEXT NOT NULL
-        )
-      `);
-
-      type KvRow = { key: string; value: string };
-      type KvInput = { key: string; value: string };
-      const kvTable = defineTable<KvRow, KvInput>({
-        name: "kv_store",
-        primaryKey: "key",
-        schema: {
-          key: col.simple<string>(),
-          value: col.simple<string>(),
-        },
-      });
-
-      const row = await kvTable.insert({
-        key: "test-key",
-        value: "test-value",
-      });
-      expect(row.key).toBe("test-key");
-      expect(row.value).toBe("test-value");
-
-      // Verify it was actually stored
-      const fetched = await kvTable.findById("test-key");
-      expect(fetched).not.toBeNull();
-      expect(fetched?.value).toBe("test-value");
-    });
-  });
-
-  describe("table utilities - inputKeyMap fallback", () => {
-    test("inputKeyMap maps single-word columns to themselves", async () => {
-      const { buildInputKeyMap } = await import("#lib/db/table.ts");
-      // Single-word columns like "name" should map to themselves
-      // This exercises the ?? dbCol fallback since toCamelCase("name") === "name"
-      const map = buildInputKeyMap(["name", "max_attendees"]);
-      expect(map.name).toBe("name");
-      expect(map.max_attendees).toBe("maxAttendees");
-    });
-
-    test("getProvidedColumns uses inputKeyMap fallback for single-word keys", async () => {
-      // The events table has a 'slug' column - toCamelCase("slug") === "slug"
-      // This tests the ?? dbCol path in getInputValue and getProvidedColumns
-      const event = await createTestEvent({
-        name: "Fallback Test",
-        maxAttendees: 10,
-        thankYouUrl: "https://example.com",
-      });
-
-      // Update with only slug - exercises getInputValue with dbCol fallback
-      const updated = await eventsTable.update(event.id, {
-        name: "Updated Name",
-        slug: "updated-slug",
-        slugIndex: "updated-index",
-        maxAttendees: 10,
-        thankYouUrl: "https://example.com",
-      });
-      expect(updated?.name).toBe("Updated Name");
-    });
-  });
-
-  describe("table findAll", () => {
-    test("returns all rows from table", async () => {
-      await createTestEvent({ name: "FindAll One", maxAttendees: 10 });
-      await createTestEvent({ name: "FindAll Two", maxAttendees: 20 });
-
-      const events = await eventsTable.findAll();
-      expect(events.length).toBeGreaterThanOrEqual(2);
-      // Verify names are decrypted (read transforms applied)
-      const names = events.map((e) => e.name);
-      expect(names).toContain("FindAll One");
-      expect(names).toContain("FindAll Two");
-    });
-  });
-
-  describe("table update returns null for non-existent id", () => {
-    test("update returns null when row does not exist", async () => {
-      const result = await eventsTable.update(99999, {
-        name: "Nonexistent",
-        slug: "nonexistent",
-        slugIndex: "idx",
-        maxAttendees: 10,
-      });
-      expect(result).toBeNull();
-    });
-  });
-
-  describe("getEventsBySlugsBatch returns null for non-existent slugs", () => {
-    test("returns null entries for slugs not in database", async () => {
-      const event = await createTestEvent({
-        name: "Batch Exists",
-        maxAttendees: 10,
-      });
-      const results = await getEventsBySlugsBatch([
-        event.slug,
-        "nonexistent-slug",
-      ]);
-      expect(results.length).toBe(2);
-      expect(results[0]?.name).toBe("Batch Exists");
-      expect(results[1]).toBeNull();
-    });
-
-    test("returns empty array for empty slug input", async () => {
-      const results = await getEventsBySlugsBatch([]);
-      expect(results).toEqual([]);
-    });
-  });
-
-  describe("decryptAttendeeOrNull", () => {
-    test("returns null when row is null", async () => {
-      const { decryptAttendeeOrNull } = await import("#lib/db/attendees.ts");
-      const privateKey = await getTestPrivateKey();
-      const result = await decryptAttendeeOrNull(null, privateKey);
-      expect(result).toBeNull();
-    });
-  });
-
-  describe("updateEventLink", () => {
-    test("updates quantity with capacity guard", async () => {
-      const { updateEventLink } = await import("#lib/db/attendees.ts");
-      const event = await createTestEvent({ maxAttendees: 5 });
-      const result = await createAttendeeAtomic({
-        name: "Link",
-        email: "link@test.com",
-        bookings: [{ eventId: event.id, quantity: 2 }],
-      });
-      expect(result.success).toBe(true);
-      if (!result.success) return;
-
-      const update = await updateEventLink(result.attendees[0]!.id, event.id, {
-        quantity: 3,
-        date: null,
-      });
-      expect(update.success).toBe(true);
-
-      const raw = await getAttendeesRaw(event.id);
-      expect(raw[0]!.quantity).toBe(3);
-    });
-
-    test("rejects update that would exceed capacity", async () => {
-      const { updateEventLink } = await import("#lib/db/attendees.ts");
-      const event = await createTestEvent({ maxAttendees: 3 });
-      const result = await createAttendeeAtomic({
-        name: "Cap",
-        email: "cap@test.com",
-        bookings: [{ eventId: event.id, quantity: 2 }],
-      });
-      expect(result.success).toBe(true);
-      if (!result.success) return;
-
-      const update = await updateEventLink(result.attendees[0]!.id, event.id, {
-        quantity: 4,
-        date: null,
-      });
-      expect(update.success).toBe(false);
-    });
-
-    test("updates date for daily event link", async () => {
-      const { updateEventLink } = await import("#lib/db/attendees.ts");
-      const event = await createTestEvent({
-        maxAttendees: 10,
-        eventType: "daily",
-      });
-      const result = await createAttendeeAtomic({
-        name: "Daily",
-        email: "daily@test.com",
-        bookings: [{ eventId: event.id, date: "2026-04-07" }],
-      });
-      expect(result.success).toBe(true);
-      if (!result.success) return;
-
-      const update = await updateEventLink(result.attendees[0]!.id, event.id, {
-        quantity: 1,
-        date: "2026-04-08",
-      });
-      expect(update.success).toBe(true);
-
-      const raw = await getAttendeesRaw(event.id);
-      expect(raw[0]!.date).toBe("2026-04-08");
-    });
-  });
-
-  describe("updateCheckedIn", () => {
-    test("updates checked_in to true for existing attendee", async () => {
-      const event = await createTestEvent({ maxAttendees: 100 });
-      const attendee = await createTestAttendee(
-        event.id,
-        event.slug,
-        "Check User",
-        "check@example.com",
-      );
-
-      await updateCheckedIn(attendee.id, event.id, true);
-
-      const privateKey = await getTestPrivateKey();
-      const rows = await getAttendeesRaw(event.id);
-      const decrypted = await decryptAttendees(rows, privateKey);
-      expect(decrypted[0]?.checked_in).toBe(true);
-    });
-
-    test("updates checked_in back to false", async () => {
-      const event = await createTestEvent({ maxAttendees: 100 });
-      const attendee = await createTestAttendee(
-        event.id,
-        event.slug,
-        "Check User",
-        "check@example.com",
-      );
-
-      await updateCheckedIn(attendee.id, event.id, true);
-      await updateCheckedIn(attendee.id, event.id, false);
-
-      const privateKey = await getTestPrivateKey();
-      const rows = await getAttendeesRaw(event.id);
-      const decrypted = await decryptAttendees(rows, privateKey);
-      expect(decrypted[0]?.checked_in).toBe(false);
-    });
-  });
-
-  describe("decryptAttendees with empty checked_in", () => {
-    test("treats empty checked_in as false for pre-migration attendees", async () => {
-      const event = await createTestEvent({ maxAttendees: 100 });
-      await createTestAttendee(
-        event.id,
-        event.slug,
-        "Old User",
-        "old@example.com",
-      );
-
-      // Simulate pre-migration state: set checked_in to empty string directly
-      await getDb().execute({
-        sql: `UPDATE attendees
-              SET checked_in = ''
-              WHERE id IN (
-                SELECT attendee_id
-                FROM event_attendees
-                WHERE event_id = ?
-              )`,
-        args: [event.id],
-      });
-
-      const privateKey = await getTestPrivateKey();
-      const rows = await getAttendeesRaw(event.id);
-      // checked_in is 0 (not checked in) — raw integer from SQL before decryption
-      expect((rows[0] as unknown as Record<string, unknown>).checked_in).toBe(
-        0,
-      );
-
-      const decrypted = await decryptAttendees(rows, privateKey);
-      expect(decrypted[0]?.checked_in).toBe(false);
-    });
-  });
-
-  describe("writeClosesAt", () => {
-    test("encrypts empty string for no deadline", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const result = await writeClosesAt("");
-      expect(typeof result).toBe("string");
-      expect(result).not.toBe("");
-      const decrypted = await decrypt(result as unknown as string);
-      expect(decrypted).toBe("");
-    });
-
-    test("encrypts null as empty string", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const result = await writeClosesAt(null);
-      const decrypted = await decrypt(result as unknown as string);
-      expect(decrypted).toBe("");
-    });
-
-    test("normalizes datetime-local string without timezone as UTC", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const input = "2099-06-15T14:30";
-      const result = await writeClosesAt(input);
-      const decrypted = await decrypt(result as unknown as string);
-      expect(decrypted).toBe(new Date(`${input}Z`).toISOString());
-    });
-
-    test("handles already-normalized ISO string", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const result = await writeClosesAt("2099-06-15T14:30:00.000Z");
-      const decrypted = await decrypt(result as unknown as string);
-      expect(decrypted).toBe("2099-06-15T14:30:00.000Z");
-    });
-
-    test("normalizes timezone offset to UTC", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const input = "2099-06-15T14:30:00-05:00";
-      const result = await writeClosesAt(input);
-      const decrypted = await decrypt(result as unknown as string);
-      expect(decrypted).toBe(new Date(input).toISOString());
-    });
-  });
-
-  describe("writeEventDate", () => {
-    test("encrypts empty string for no date", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const result = await writeEventDate("");
-      expect(typeof result).toBe("string");
-      expect(result).not.toBe("");
-      const decrypted = await decrypt(result);
-      expect(decrypted).toBe("");
-    });
-
-    test("normalizes datetime-local string without timezone as UTC", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const input = "2026-06-15T14:00";
-      const result = await writeEventDate(input);
-      const decrypted = await decrypt(result);
-      expect(decrypted).toBe(new Date(`${input}Z`).toISOString());
-    });
-
-    test("handles already-normalized ISO string", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const result = await writeEventDate("2026-06-15T14:00:00.000Z");
-      const decrypted = await decrypt(result);
-      expect(decrypted).toBe("2026-06-15T14:00:00.000Z");
-    });
-
-    test("normalizes timezone offset to UTC", async () => {
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const input = "2026-06-15T14:00:00+02:00";
-      const result = await writeEventDate(input);
-      const decrypted = await decrypt(result);
-      expect(decrypted).toBe(new Date(input).toISOString());
-    });
-
-    test("returns empty string for invalid datetime", async () => {
-      const errorSpy = spy(console, "error");
-      const { decrypt } = await import("#lib/crypto/encryption.ts");
-      const result = await writeEventDate("not-a-dateZ");
-      const decrypted = await decrypt(result);
-      expect(decrypted).toBe("");
-      expect(errorSpy.calls.length).toBeGreaterThan(0);
-      errorSpy.restore();
-    });
-  });
-
-  describe("event date read transform", () => {
-    test("returns empty string for no-date event", async () => {
-      const event = await eventsTable.insert({
-        name: "test",
-        slug: "test-date-read-1",
-        slugIndex: await computeSlugIndex("test-date-read-1"),
-        maxAttendees: 100,
-        maxPrice: 10000,
-        date: "",
-      });
-      const saved = await getEventWithCount(event.id);
-      expect(saved?.date).toBe("");
-    });
-
-    test("returns normalized ISO string for valid datetime", async () => {
-      const event = await eventsTable.insert({
-        name: "test",
-        slug: "test-date-read-2",
-        slugIndex: await computeSlugIndex("test-date-read-2"),
-        maxAttendees: 100,
-        maxPrice: 10000,
-        date: "2026-06-15T14:00",
-      });
-      const saved = await getEventWithCount(event.id);
-      expect(saved?.date).toBe("2026-06-15T14:00:00.000Z");
-    });
-  });
-
-  describe("closes_at read transform", () => {
-    test("returns null for no-deadline event", async () => {
-      const event = await eventsTable.insert({
-        name: "test",
-        slug: "test-read-1",
-        slugIndex: await computeSlugIndex("test-read-1"),
-        maxAttendees: 100,
-        maxPrice: 10000,
-        closesAt: "",
-      });
-      const saved = await getEventWithCount(event.id);
-      expect(saved?.closes_at).toBeNull();
-    });
-
-    test("returns normalized ISO string for valid datetime", async () => {
-      const event = await eventsTable.insert({
-        name: "test",
-        slug: "test-read-2",
-        slugIndex: await computeSlugIndex("test-read-2"),
-        maxAttendees: 100,
-        maxPrice: 10000,
-        closesAt: "2099-12-31T23:59",
-      });
-      const saved = await getEventWithCount(event.id);
-      expect(saved?.closes_at).toBe("2099-12-31T23:59:00.000Z");
-    });
-  });
-
-  describe("bookable_days read transform", () => {
-    test("returns empty array when DB contains non-array JSON", async () => {
-      const event = await eventsTable.insert({
-        name: "test-bd",
-        slug: "test-bd-1",
-        slugIndex: await computeSlugIndex("test-bd-1"),
-        maxAttendees: 100,
-        maxPrice: 10000,
-      });
-      // Overwrite bookable_days with non-array JSON directly in DB
-      await getDb().execute({
-        sql: "UPDATE events SET bookable_days = ? WHERE id = ?",
-        args: ['"not-an-array"', event.id],
-      });
-      const saved = await getEventWithCount(event.id);
-      expect(saved?.bookable_days).toEqual([]);
-    });
-  });
-
-  describe("groups", () => {
-    test("groupsTable CRUD works", async () => {
-      const created = await createTestGroup({
-        name: "DB Group",
-        slug: "db-group",
-      });
-
-      const fetched = await groupsTable.findById(created.id);
-      expect(fetched).not.toBeNull();
-      expect(fetched?.name).toBe("DB Group");
-      expect(fetched?.slug).toBe("db-group");
-
-      const updated = await groupsTable.update(created.id, {
-        name: "DB Group Updated",
-        termsAndConditions: "Terms",
-      });
-      expect(updated?.name).toBe("DB Group Updated");
-      expect(updated?.terms_and_conditions).toBe("Terms");
-
-      await groupsTable.deleteById(created.id);
-      expect(await groupsTable.findById(created.id)).toBeNull();
-    });
-
-    test("getAllGroups returns decrypted groups ordered by id", async () => {
-      const g1 = await createTestGroup({ name: "Group A", slug: "group-a" });
-      const g2 = await createTestGroup({ name: "Group B", slug: "group-b" });
-      const groups = await getAllGroups();
-      expect(groups.length).toBe(2);
-      expect(groups[0]?.id).toBe(g1.id);
-      expect(groups[1]?.id).toBe(g2.id);
-      expect(groups[0]?.name).toBe("Group A");
-      expect(groups[1]?.name).toBe("Group B");
-    });
-
-    test("getGroupBySlugIndex returns group or null", async () => {
-      const group = await createTestGroup({
-        name: "Index Group",
-        slug: "idx-group",
-      });
-
-      const found = await getGroupBySlugIndex(
-        await computeGroupSlugIndex("idx-group"),
-      );
-      expect(found?.slug).toBe(group.slug);
-      expect(await getGroupBySlugIndex("missing")).toBeNull();
-    });
-
-    test("isGroupSlugTaken checks both groups and events", async () => {
-      const groupSlug = "taken-by-group";
-      const created = await createTestGroup({
-        name: "Taken",
-        slug: groupSlug,
-      });
-
-      expect(await isGroupSlugTaken(groupSlug)).toBe(true);
-      expect(await isGroupSlugTaken(groupSlug, created.id)).toBe(false);
-
-      const event = await createTestEvent({ name: "Taken Event" });
-      expect(await isGroupSlugTaken(event.slug)).toBe(true);
-    });
-
-    test("getActiveEventsByGroupId returns active events with attendee counts", async () => {
-      const group = await createTestGroup({
-        name: "Events Group",
-        slug: "events-group",
-      });
-
-      const e1 = await createTestEvent({
-        name: "Active In Group",
-        maxAttendees: 10,
-        groupId: group.id,
-      });
-      const e2 = await createTestEvent({
-        name: "Inactive In Group",
-        maxAttendees: 10,
-        groupId: group.id,
-      });
-      await getDb().execute({
-        sql: "UPDATE events SET active = 0 WHERE id = ?",
-        args: [e2.id],
-      });
-
-      const attendee = await createAttendeeAtomic({
-        name: "A",
-        email: "a@example.com",
-        bookings: [{ eventId: e1.id, quantity: 3 }],
-      });
-      if (!attendee.success) throw new Error("Failed to create attendee");
-
-      const events = await getActiveEventsByGroupId(group.id);
-      expect(events.length).toBe(1);
-      expect(events[0]?.id).toBe(e1.id);
-      expect(events[0]?.attendee_count).toBe(3);
-    });
-
-    test("resetGroupEvents sets group_id to 0", async () => {
-      const group = await createTestGroup({
-        name: "Reset Group",
-        slug: "reset-group",
-      });
-      const event = await createTestEvent({
-        name: "Reset Event",
-        groupId: group.id,
-        maxAttendees: 10,
-      });
-      await resetGroupEvents(group.id);
-      expect((await getEvent(event.id))?.group_id).toBe(0);
-    });
-
-    /** Create a capped group with two events (each with event-level max of 10) */
-    const createCappedGroupWithEvents = async (
-      groupMax: number,
-      slug: string,
-      overrides?: { eventType?: "standard" | "daily" },
-    ) => {
-      const group = await createTestGroup({
-        name: slug,
-        slug,
-        maxAttendees: groupMax,
-      });
-      const e1 = await createTestEvent({
-        name: `${slug}-a`,
-        maxAttendees: 10,
-        groupId: group.id,
-        eventType: overrides?.eventType,
-      });
-      const e2 = await createTestEvent({
-        name: `${slug}-b`,
-        maxAttendees: 10,
-        groupId: group.id,
-        eventType: overrides?.eventType,
-      });
-      return { group, e1, e2 };
-    };
-
-    /** Book attendees atomically with minimal boilerplate */
-    const book = (eventId: number, quantity: number, date?: string) =>
-      createAttendeeAtomic({
-        name: `attendee-${eventId}-${quantity}`,
-        email: `a${eventId}q${quantity}@example.com`,
-        bookings: [{ eventId, quantity, date }],
-      });
-
-    test("createAttendeeAtomic enforces group max_attendees across events", async () => {
-      const { e1, e2 } = await createCappedGroupWithEvents(5, "capped");
-
-      // Book 3 on event A — should succeed (group total: 3/5)
-      expect((await book(e1.id, 3)).success).toBe(true);
-
-      // Book 3 on event B — should fail (group total would be 6/5)
-      const r2 = await book(e2.id, 3);
-      expect(r2.success).toBe(false);
-      if (!r2.success) expect(r2.reason).toBe("capacity_exceeded");
-
-      // Book 2 on event B — should succeed (group total: 5/5)
-      expect((await book(e2.id, 2)).success).toBe(true);
-    });
-
-    test("createAttendeeAtomic allows booking when group has no max (0)", async () => {
-      const group = await createTestGroup({
-        name: "unlimited",
-        slug: "unlimited",
-      });
-      const event = await createTestEvent({
-        name: "unlimited-event",
-        maxAttendees: 100,
-        groupId: group.id,
-      });
-
-      expect((await book(event.id, 50)).success).toBe(true);
-    });
-
-    test("hasAvailableSpots checks group capacity", async () => {
-      const { hasAvailableSpots } = await import("#lib/db/attendees.ts");
-      const { e1, e2 } = await createCappedGroupWithEvents(3, "spots");
-
-      await book(e1.id, 2);
-
-      // Event B has plenty of room, but group is almost full
-      expect(await hasAvailableSpots(e2.id, 1)).toBe(true);
-      expect(await hasAvailableSpots(e2.id, 2)).toBe(false);
-    });
-
-    test("checkBatchAvailability checks group capacity", async () => {
-      const { checkBatchAvailability } = await import("#lib/db/attendees.ts");
-      const { e1, e2 } = await createCappedGroupWithEvents(4, "batch");
-
-      // Both within event limits but combined exceeds group limit
-      expect(
-        await checkBatchAvailability([
-          { eventId: e1.id, quantity: 3 },
-          { eventId: e2.id, quantity: 2 },
-        ]),
-      ).toBe(false);
-
-      // Both within event and group limits
-      expect(
-        await checkBatchAvailability([
-          { eventId: e1.id, quantity: 2 },
-          { eventId: e2.id, quantity: 2 },
-        ]),
-      ).toBe(true);
-    });
-
-    test("checkBatchAvailability skips group check when group has no limit", async () => {
-      const { checkBatchAvailability } = await import("#lib/db/attendees.ts");
-      const group = await createTestGroup({
-        name: "no-limit",
-        slug: "no-limit",
-      });
-      const e1 = await createTestEvent({
-        name: "no-limit-a",
-        maxAttendees: 100,
-        groupId: group.id,
-      });
-      const ungrouped = await createTestEvent({
-        name: "ungrouped",
-        maxAttendees: 100,
-      });
-
-      // Mix of grouped (no limit) and ungrouped events
-      expect(
-        await checkBatchAvailability([
-          { eventId: e1.id, quantity: 50 },
-          { eventId: ungrouped.id, quantity: 50 },
-        ]),
-      ).toBe(true);
-    });
-
-    test("checkBatchAvailability handles events from multiple groups", async () => {
-      const { checkBatchAvailability } = await import("#lib/db/attendees.ts");
-      const { e1: eA } = await createCappedGroupWithEvents(3, "multi-a");
-      const { e1: eB } = await createCappedGroupWithEvents(3, "multi-b");
-
-      // Each group can hold 3, so 2+2 across different groups is fine
-      expect(
-        await checkBatchAvailability([
-          { eventId: eA.id, quantity: 2 },
-          { eventId: eB.id, quantity: 2 },
-        ]),
-      ).toBe(true);
-    });
-
-    test("group capacity check handles deleted group gracefully", async () => {
-      const { hasAvailableSpots } = await import("#lib/db/attendees.ts");
-      const group = await createTestGroup({
-        name: "delete-me",
-        slug: "delete-me",
-        maxAttendees: 5,
-      });
-      const event = await createTestEvent({
-        name: "orphan-event",
-        maxAttendees: 10,
-        groupId: group.id,
-      });
-      // Delete the group but leave event pointing to it
-      const { groupsTable } = await import("#lib/db/groups.ts");
-      await groupsTable.deleteById(group.id);
-
-      // Should still work — getGroupMaxAttendees returns 0 for missing group
-      expect(await hasAvailableSpots(event.id, 1)).toBe(true);
-    });
-
-    test("group max_attendees is per-date for daily events", async () => {
-      const { e1: event } = await createCappedGroupWithEvents(3, "daily", {
-        eventType: "daily",
-      });
-
-      // Book 3 for date A — fills group for that date
-      expect((await book(event.id, 3, "2026-07-01")).success).toBe(true);
-
-      // Book 1 for date A — should fail (group full for date A)
-      expect((await book(event.id, 1, "2026-07-01")).success).toBe(false);
-
-      // Book 3 for date B — should succeed (different date)
-      expect((await book(event.id, 3, "2026-07-02")).success).toBe(true);
-    });
-
-    test("daily group cap counts across multiple events for same date", async () => {
-      const { e1, e2 } = await createCappedGroupWithEvents(4, "daily-multi", {
-        eventType: "daily",
-      });
-
-      // Book 2 on event A for date X (group total: 2/4)
-      expect((await book(e1.id, 2, "2026-07-01")).success).toBe(true);
-
-      // Book 2 on event B for same date X (group total: 4/4)
-      expect((await book(e2.id, 2, "2026-07-01")).success).toBe(true);
-
-      // Book 1 more on event B for same date — should fail (group full for this date)
-      expect((await book(e2.id, 1, "2026-07-01")).success).toBe(false);
-
-      // Book on event B for different date — should succeed (separate date budget)
-      expect((await book(e2.id, 3, "2026-07-02")).success).toBe(true);
-    });
-
-    test("hasAvailableSpots checks group capacity for daily events with date", async () => {
-      const { hasAvailableSpots } = await import("#lib/db/attendees.ts");
-      const { e1, e2 } = await createCappedGroupWithEvents(3, "daily-spots", {
-        eventType: "daily",
-      });
-
-      // Book 2 on event A for a specific date
-      await book(e1.id, 2, "2026-08-01");
-
-      // Event B has its own capacity (10), but group only has 1 spot left for this date
-      expect(await hasAvailableSpots(e2.id, 1, "2026-08-01")).toBe(true);
-      expect(await hasAvailableSpots(e2.id, 2, "2026-08-01")).toBe(false);
-
-      // Different date should have full group budget available
-      expect(await hasAvailableSpots(e2.id, 3, "2026-08-02")).toBe(true);
-    });
-
-    test("checkBatchAvailability checks group capacity with date for daily events", async () => {
-      const { checkBatchAvailability } = await import("#lib/db/attendees.ts");
-      const { e1, e2 } = await createCappedGroupWithEvents(5, "daily-batch", {
-        eventType: "daily",
-      });
-
-      // Both events in same group — combined quantity exceeds group cap for this date
-      expect(
-        await checkBatchAvailability(
-          [
-            { eventId: e1.id, quantity: 3 },
-            { eventId: e2.id, quantity: 3 },
-          ],
-          "2026-09-01",
-        ),
-      ).toBe(false);
-
-      // Within group cap for this date
-      expect(
-        await checkBatchAvailability(
-          [
-            { eventId: e1.id, quantity: 2 },
-            { eventId: e2.id, quantity: 3 },
-          ],
-          "2026-09-01",
-        ),
-      ).toBe(true);
-    });
-
-    test("checkBatchAvailability considers pre-existing attendees in group", async () => {
-      const { checkBatchAvailability } = await import("#lib/db/attendees.ts");
-      const { e1, e2 } = await createCappedGroupWithEvents(5, "pre-exist");
-
-      // Book 3 on event A first
-      await book(e1.id, 3);
-
-      // Batch check: 3 more would exceed group cap (3+3=6 > 5)
-      expect(
-        await checkBatchAvailability([{ eventId: e2.id, quantity: 3 }]),
-      ).toBe(false);
-
-      // 2 more fits exactly (3+2=5)
-      expect(
-        await checkBatchAvailability([{ eventId: e2.id, quantity: 2 }]),
-      ).toBe(true);
-    });
-
-    test("event-level cap rejects even when group has room", async () => {
-      const group = await createTestGroup({
-        name: "big-group",
-        slug: "big-group",
-        maxAttendees: 100,
-      });
-      const event = await createTestEvent({
-        name: "small-event",
-        maxAttendees: 2,
-        groupId: group.id,
-      });
-
-      // Event max is 2, group max is 100 — event cap should be the constraint
-      expect((await book(event.id, 2)).success).toBe(true);
-      const r = await book(event.id, 1);
-      expect(r.success).toBe(false);
-      if (!r.success) expect(r.reason).toBe("capacity_exceeded");
-    });
-
-    test("hasAvailableSpots respects event cap even when group has room", async () => {
-      const { hasAvailableSpots } = await import("#lib/db/attendees.ts");
-      const group = await createTestGroup({
-        name: "big-group2",
-        slug: "big-group2",
-        maxAttendees: 100,
-      });
-      const event = await createTestEvent({
-        name: "tiny-event",
-        maxAttendees: 1,
-        groupId: group.id,
-      });
-
-      await book(event.id, 1);
-      expect(await hasAvailableSpots(event.id, 1)).toBe(false);
-    });
-
-    test("checkBatchAvailability rejects when one group is full and another has room", async () => {
-      const { checkBatchAvailability } = await import("#lib/db/attendees.ts");
-      const { e1: fullGroupEvent } = await createCappedGroupWithEvents(
-        2,
-        "full-grp",
-      );
-      const { e1: openGroupEvent } = await createCappedGroupWithEvents(
-        10,
-        "open-grp",
-      );
-
-      // Fill the first group
-      await book(fullGroupEvent.id, 2);
-
-      // Batch with one event from full group + one from open group
-      expect(
-        await checkBatchAvailability([
-          { eventId: fullGroupEvent.id, quantity: 1 },
-          { eventId: openGroupEvent.id, quantity: 1 },
-        ]),
-      ).toBe(false);
-
-      // Just the open group event should work
-      expect(
-        await checkBatchAvailability([
-          { eventId: openGroupEvent.id, quantity: 1 },
-        ]),
-      ).toBe(true);
-    });
-  });
-});
-
-/**
- * Migration test: verifies that migrating from the main-branch schema
- * (attendees with event_id FK) to the new schema (event_attendees table)
- * works correctly even when PRAGMA foreign_keys=OFF is ineffective
- * (as happens in remote libsql / Turso where it doesn't persist across
- * HTTP requests).
- *
- * The legacy schema has:
- *   attendees.event_id → FOREIGN KEY REFERENCES events(id)
- *   processed_payments.attendee_id → FOREIGN KEY REFERENCES attendees(id)
- *   attendee_answers.attendee_id → FOREIGN KEY REFERENCES attendees(id)
- *
- * The migration needs to recreate the attendees table (to drop event_id,
- * date, quantity columns). DROP TABLE attendees fails with FK constraint
- * when processed_payments or attendee_answers have data referencing it.
- */
-describe("event_attendees migration from legacy schema", () => {
-  afterEach(() => {
-    resetDb();
-  });
-
-  /** SQL statements that create the complete legacy schema (as on main) */
-  const LEGACY_SCHEMA_SQL = [
-    "CREATE TABLE settings (key TEXT PRIMARY KEY, value TEXT NOT NULL)",
-    `CREATE TABLE events (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      created TEXT NOT NULL,
-      max_attendees INTEGER NOT NULL,
-      thank_you_url TEXT,
-      unit_price INTEGER,
-      max_quantity INTEGER NOT NULL DEFAULT 1,
-      webhook_url TEXT,
-      slug TEXT,
-      slug_index TEXT,
-      group_id INTEGER NOT NULL DEFAULT 0,
-      active INTEGER NOT NULL DEFAULT 1,
-      fields TEXT NOT NULL DEFAULT 'email',
-      closes_at TEXT,
-      name TEXT NOT NULL DEFAULT '',
-      description TEXT NOT NULL DEFAULT '',
-      event_type TEXT NOT NULL DEFAULT 'standard',
-      bookable_days TEXT NOT NULL DEFAULT '["Monday","Tuesday","Wednesday","Thursday","Friday","Saturday","Sunday"]',
-      minimum_days_before INTEGER NOT NULL DEFAULT 1,
-      maximum_days_after INTEGER NOT NULL DEFAULT 90,
-      date TEXT NOT NULL DEFAULT '',
-      location TEXT NOT NULL DEFAULT '',
-      image_url TEXT NOT NULL DEFAULT '',
-      attachment_url TEXT NOT NULL DEFAULT '',
-      attachment_name TEXT NOT NULL DEFAULT '',
-      non_transferable INTEGER NOT NULL DEFAULT 0,
-      can_pay_more INTEGER NOT NULL DEFAULT 0,
-      hidden INTEGER NOT NULL DEFAULT 0,
-      max_price INTEGER NOT NULL DEFAULT 0
-    )`,
-    "CREATE UNIQUE INDEX idx_events_slug_index ON events(slug_index)",
-    `CREATE TABLE users (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      username_hash TEXT NOT NULL,
-      username_index TEXT NOT NULL,
-      password_hash TEXT NOT NULL DEFAULT '',
-      wrapped_data_key TEXT,
-      admin_level TEXT NOT NULL,
-      invite_code_hash TEXT,
-      invite_expiry TEXT
-    )`,
-    "CREATE UNIQUE INDEX idx_users_username_index ON users(username_index)",
-    `CREATE TABLE sessions (
-      token TEXT PRIMARY KEY,
-      csrf_token TEXT NOT NULL,
-      expires INTEGER NOT NULL,
-      wrapped_data_key TEXT,
-      user_id INTEGER
-    )`,
-    `CREATE TABLE login_attempts (
-      ip TEXT PRIMARY KEY,
-      attempts INTEGER NOT NULL DEFAULT 0,
-      locked_until INTEGER
-    )`,
-    `CREATE TABLE attendees (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      event_id INTEGER NOT NULL,
-      name TEXT NOT NULL,
-      email TEXT NOT NULL,
-      created TEXT NOT NULL,
-      payment_id TEXT,
-      quantity INTEGER NOT NULL DEFAULT 1,
-      phone TEXT NOT NULL DEFAULT '',
-      ticket_token TEXT NOT NULL DEFAULT '',
-      price_paid TEXT,
-      checked_in TEXT NOT NULL DEFAULT '',
-      date TEXT DEFAULT NULL,
-      address TEXT NOT NULL DEFAULT '',
-      special_instructions TEXT NOT NULL DEFAULT '',
-      ticket_token_index TEXT,
-      refunded TEXT NOT NULL DEFAULT '',
-      attachment_downloads INTEGER NOT NULL DEFAULT 0,
-      pii_blob TEXT NOT NULL DEFAULT '',
-      checked_in_v2 INTEGER NOT NULL DEFAULT 0,
-      refunded_v2 INTEGER NOT NULL DEFAULT 0,
-      price_paid_v2 INTEGER NOT NULL DEFAULT 0,
-      FOREIGN KEY (event_id) REFERENCES events(id)
-    )`,
-    `CREATE UNIQUE INDEX
-     idx_attendees_ticket_token_index
-     ON attendees(ticket_token_index)`,
-    `CREATE TABLE processed_payments (
-      payment_session_id TEXT PRIMARY KEY,
-      attendee_id INTEGER,
-      processed_at TEXT NOT NULL,
-      ticket_tokens TEXT NOT NULL DEFAULT '',
-      FOREIGN KEY (attendee_id) REFERENCES attendees(id)
-    )`,
-    `CREATE TABLE activity_log (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      created TEXT NOT NULL,
-      event_id INTEGER,
-      message TEXT NOT NULL,
-      FOREIGN KEY (event_id) REFERENCES events(id) ON DELETE SET NULL
-    )`,
-    `CREATE TABLE groups (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      slug TEXT NOT NULL,
-      slug_index TEXT NOT NULL,
-      name TEXT NOT NULL,
-      terms_and_conditions TEXT NOT NULL DEFAULT '',
-      max_attendees INTEGER NOT NULL DEFAULT 0
-    )`,
-    "CREATE UNIQUE INDEX idx_groups_slug_index ON groups(slug_index)",
-    `CREATE TABLE holidays (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      name TEXT NOT NULL,
-      start_date TEXT NOT NULL,
-      end_date TEXT NOT NULL
-    )`,
-    `CREATE TABLE api_keys (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      user_id INTEGER NOT NULL,
-      key_index TEXT NOT NULL,
-      wrapped_data_key TEXT NOT NULL,
-      name TEXT NOT NULL,
-      created TEXT NOT NULL,
-      last_used TEXT NOT NULL DEFAULT '',
-      FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
-    )`,
-    "CREATE UNIQUE INDEX idx_api_keys_key_index ON api_keys(key_index)",
-    `CREATE TABLE questions (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      text TEXT NOT NULL
-    )`,
-    `CREATE TABLE answers (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      question_id INTEGER NOT NULL,
-      text TEXT NOT NULL,
-      sort_order INTEGER NOT NULL DEFAULT 0,
-      FOREIGN KEY (question_id) REFERENCES questions(id)
-    )`,
-    "CREATE INDEX idx_answers_question_id ON answers(question_id)",
-    `CREATE TABLE event_questions (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      event_id INTEGER NOT NULL,
-      question_id INTEGER NOT NULL,
-      sort_order INTEGER NOT NULL DEFAULT 0,
-      FOREIGN KEY (event_id) REFERENCES events(id),
-      FOREIGN KEY (question_id) REFERENCES questions(id)
-    )`,
-    "CREATE INDEX idx_event_questions_event_id ON event_questions(event_id)",
-    `CREATE UNIQUE INDEX
-     idx_event_questions_unique
-     ON event_questions(event_id, question_id)`,
-    `CREATE TABLE built_sites (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      site_data TEXT NOT NULL,
-      created TEXT NOT NULL
-    )`,
-    `CREATE TABLE attendee_answers (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      attendee_id INTEGER NOT NULL,
-      answer_id INTEGER NOT NULL,
-      FOREIGN KEY (attendee_id) REFERENCES attendees(id),
-      FOREIGN KEY (answer_id) REFERENCES answers(id)
-    )`,
-    `CREATE INDEX idx_attendee_answers_attendee_id
-     ON attendee_answers(attendee_id)`,
-    `CREATE INDEX idx_attendee_answers_answer_id
-     ON attendee_answers(answer_id)`,
-    `CREATE UNIQUE INDEX
-     idx_attendee_answers_unique
-     ON attendee_answers(attendee_id, answer_id)`,
-  ];
-
-  /** Create the legacy schema and return the client */
-  const createLegacyDb = async () => {
-    setupTestEncryptionKey();
-    const client = createClient({ url: ":memory:" });
-    setDb(client);
-    for (const sql of LEGACY_SCHEMA_SQL) {
-      await client.execute(sql);
-    }
-    return client;
-  };
-
-  /**
-   * Stub PRAGMA foreign_keys = OFF to be a no-op.
-   * This simulates remote libsql (Turso) where PRAGMA doesn't persist
-   * across HTTP requests — the execute() and batch() calls are separate
-   * requests, so PRAGMA set in execute() doesn't affect the batch().
-   *
-   * In the in-memory client, PRAGMA carries over (single connection),
-   * so without this stub the migration would succeed and hide the bug.
-   */
-  const stubPragmaForeignKeysOff = (
-    client: ReturnType<typeof createClient>,
-  ) => {
-    const origExecute = client.execute.bind(client);
-    return stub(client, "execute", (stmt: unknown) => {
-      const sql =
-        typeof stmt === "string" ? stmt : (stmt as { sql: string }).sql;
-      if (/PRAGMA\s+foreign_keys\s*=\s*OFF/i.test(sql)) {
-        return Promise.resolve({
-          columns: [],
-          columnTypes: [],
-          rows: [],
-          rowsAffected: 0,
-          lastInsertRowid: 0n,
-          toJSON: () => ({
-            columns: [],
-            columnTypes: [],
-            rows: [],
-            rowsAffected: 0,
-            lastInsertRowid: "0",
-          }),
-        } as ResultSet);
-      }
-      return origExecute(stmt as Parameters<typeof origExecute>[0]);
-    });
-  };
-
-  test("migration backfills event_attendees and preserves processed_payments", async () => {
-    const client = await createLegacyDb();
-
-    // Insert legacy data: event with a date, attendee, and a processed payment
-    await client.execute(
-      insert("events", {
-        id: 1,
-        created: "2024-01-01T00:00:00Z",
-        max_attendees: 100,
-        name: "Test Event",
-      }),
-    );
-    await client.execute(
-      insert("attendees", {
-        id: 1,
-        event_id: 1,
-        name: "Test User",
-        email: "test@example.com",
-        created: "2024-01-01T00:00:00Z",
-        quantity: 2,
-        date: "2024-06-15",
-        checked_in_v2: 0,
-        refunded_v2: 0,
-        price_paid_v2: 1000,
-      }),
-    );
-    await client.execute(
-      insert("processed_payments", {
-        payment_session_id: "ps_test_123",
-        attendee_id: 1,
-        processed_at: "2024-01-01T00:00:00Z",
-      }),
-    );
-
-    // Simulate production: PRAGMA foreign_keys=OFF doesn't work
-    const pragmaStub = stubPragmaForeignKeysOff(client);
-    try {
-      await initDb();
-    } finally {
-      pragmaStub.restore();
-    }
-
-    // Verify attendee was migrated to event_attendees
-    const ea = await client.execute("SELECT * FROM event_attendees");
-    expect(ea.rows.length).toBe(1);
-    expect(ea.rows[0]!.event_id).toBe(1);
-    expect(ea.rows[0]!.attendee_id).toBe(1);
-    expect(ea.rows[0]!.quantity).toBe(2);
-    expect(ea.rows[0]!.price_paid).toBe(1000);
-    // date "2024-06-15" → start_at "2024-06-15T00:00:00Z", end_at "2024-06-16T00:00:00Z"
-    expect(ea.rows[0]!.start_at).toBe("2024-06-15T00:00:00Z");
-    expect(ea.rows[0]!.end_at).toBe("2024-06-16T00:00:00Z");
-
-    // Verify attendees table no longer has deprecated columns
-    const cols = await client.execute("PRAGMA table_info(attendees)");
-    const colNames = cols.rows.map((r) => r.name);
-    expect(colNames).not.toContain("event_id");
-    expect(colNames).not.toContain("date");
-    expect(colNames).not.toContain("quantity");
-    expect(colNames).toContain("id");
-    expect(colNames).toContain("pii_blob");
-
-    // Verify processed_payments data survived the migration
-    const payments = await client.execute("SELECT * FROM processed_payments");
-    expect(payments.rows.length).toBe(1);
-    expect(payments.rows[0]!.attendee_id).toBe(1);
-  });
-});
